### 2024-03-26 #6540 @editorialbot
[PRE REVIEW]: COBRAPRO: A MATLAB toolbox for Physics-based Battery Modeling and Co-simulation Parameter Optimization
**Submitting author:** @COBRAPROsimulator (<a href="http://orcid.org/0009-0005-9878-3537">Sara Ha</a>)
**Repository:** https://github.com/COBRAPROsimulator/COBRAPRO
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
