+++ Examples/Parameter_Identification_Results/DFN_pso_UDDS_validation.m
@@ -5,7 +5,7 @@
 % C/20 discharge and HPPC data.
 %
 % Parameters identified during C/20 discharge: {theta100_p,theta0_p,theta100_n,theta0_n}
-% Parameters identified during HPPC: {kn, Dsp, c0, kp, Dsn, Kappa, De, t1_constant}
+% Parameters identified during HPPC: {kp, kn, Dsp, Dsn, Kappa, De, t1_constant, c0}
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % COBRAPRO: Co-simulation Battery Modeling for Accelerated Parameter Optimization
