+++ paper/paper.md
@@ -57,23 +57,98 @@ In contrast, several open-source DFN model simulation tools have been released s
   - Constant current profiles
   - Hybrid pulse power characterization (HPPC) profiles
   - Dynamic current profiles 
-- **Local sensitivity analysis:** Perturbs parameters around specific reference values to determine sensitive parameters for a given current profile
+- **Local sensitivity analysis:** Perturbs parameters around nominal values to determine sensitive parameters for a given current profile
 
 # Examples 
-```r
-flying %>% 
-drop_summary(last_col = "location_census_region") %>% 
-print(n = Inf) 
+In the `Examples/Parameter_Identification_Routines` folder, two example codes are provided that demonstrate a two-step parameter identification process. In step 1, the stoichiometric parameters are identified in `DFN_pso_0_05C.m` using the experimentally obtained C/20 discharge profile. In step 2, the electrolyte transport and kinetic parameters are identified in `DFN_pso_HPPC.m` using the experimentally obtained HPPC profile. 
+
+In `DFN_pso_0_05C.m`, the user inputs are given as:
+```
+%% User Input  
+
+%--------------------------------------------------------------------------
+% Load nominal parameters
+%--------------------------------------------------------------------------
+param = Parameters_LG_INR21700_M50;
+
+%--------------------------------------------------------------------------
+% Enter mat file name where your PSO results will be stored
+%--------------------------------------------------------------------------
+file_name = 'pso_0_05C';
+
+%--------------------------------------------------------------------------
+% Enter names of parameters to identify (make sure names match the
+% parameter names in "param" structure containing the nominal parameters)
+%--------------------------------------------------------------------------
+param_CC = {'theta100_p', 'theta100_n', 'theta0_p', 'theta0_n'};
+
+%--------------------------------------------------------------------------
+% Enter lower and upper bounds of parameters to identify 
+%--------------------------------------------------------------------------
+%   Example for parameter name "xxx":
+%   lower_bounds.xxx = ...
+%   upper_bounds.xxx = ...
+%--------------------------------------------------------------------------
+% theta100_p
+lower_bounds.theta100_p = 0.22; 
+upper_bounds.theta100_p = 0.34;
+% theta100_n
+lower_bounds.theta100_n = 0.7; 
+upper_bounds.theta100_n = 1; 
+% theta0_p
+lower_bounds.theta0_p = 0.7; 
+upper_bounds.theta0_p = 1; 
+% theta0_n
+lower_bounds.theta0_n = 0.015; 
+upper_bounds.theta0_n = 0.04; 
+
+%--------------------------------------------------------------------------
+% Enter number of particles for PSO
+%--------------------------------------------------------------------------
+particle_num = 100;
+
+%--------------------------------------------------------------------------
+% Load Experimental Data 
+%--------------------------------------------------------------------------
+%   t: Should be a vector consisting of your time experiment data      [s] (Mx1)
+%   I: Should be a vector consisting of your current experiment data   [A] (Mx1) (negative current: discharging)
+%   V: Should be a vector consisting of your volatge experiemntal data [V] (Mx1)
+%   -> where M is the total number of data points in your experiment
+%--------------------------------------------------------------------------
+% C/20 capacity test conducted on LG INR21700 M50T cells
+load('data_INR21700_M50T/capacity_test_data_W8_Diag1.mat')
+
+t = t_data;
+I = I_data;
+V = V_data;
+
+%--------------------------------------------------------------------------
+% Choose the desired sampling time [s] to resample your experiment data 
+% (the interpolated experimental data will be used in the PSO to compute the objective function)
+%--------------------------------------------------------------------------
+deltaT_exp = 1; 
+
[…]
