+++ paper/paper.md
@@ -62,8 +62,8 @@ In contrast, several open-source DFN model simulation tools have been released s
 # Examples 
 In the `Examples/Parameter_Identification_Routines` folder, two example codes are provided that demonstrate a two-step parameter identification process. In step 1, the stoichiometric parameters are identified in `DFN_pso_0_05C.m` using the experimentally obtained C/20 discharge profile. In step 2, the electrolyte transport and kinetic parameters are identified in `DFN_pso_HPPC.m` using the experimentally obtained HPPC profile. 
 
-## C/20 Discharge Identification (`DFN_pso_0_05C.m`)
-The code is written such that users only need to modify the `User Input` section in the code. In the `User Input` section, the parameters to be identified, the upper and lower bounds for each parameter, and the experimental data are defined:
+## C/20 Discharge Identification
+`Examples/Parameter_Identification_Routines/DFN_pso_0_05C.m` is written such that users only need to modify the `User Input`. The names of the parameters to be identified, upper and lower bounds for each parameter, and the experimental data are defined:
 ```MATLAB
 %% User Input  
 % Load nominal parameters 
@@ -108,19 +108,8 @@ V = V_data;
 
 % Enter experimental data initial SOC [-] 
 SOC_init = 1;  
-
-% Simulation time 
-%--------------------------------------------------------------------------
-% NOTE: Simulation will exit automatically when the lower or upper cut-off
-% voltages are reached. To simulate till the cut-off voltage condition,
-% keep tf large enough, e.g., for C/20 the tf should be greater than 7200.
-%--------------------------------------------------------------------------
-% Initial simulation time [s]
-t0 = 0;
-% Final simulation time [s] 
-tf = 1e6;
 ```
-Note that the `Parameters_LG_INR21700_M50.m` function defines the nominal parameters of your cell and the simulation settings, e.g., discretization method, DAE initialization method, current type, etc.
+Note that `Parameters_LG_INR21700_M50.m` function defines the nominal parameters of your cell and the simulation settings, e.g., discretization method, DAE initialization method, current type, etc. 
 
 Once the user input has been defined, run the code to start the PSO. Once the PSO is finished, the code prints the identified parameter values and objective function values to the command window:
 ```
@@ -155,6 +144,121 @@ to the Command Window and plots the identified results as shown in \autoref{fig:
 ![C/20 discharge voltage identification results.\label{fig:V_0_05C}](voltage_0_05C_identification.png)
 
 ![C/20 discharge positive and negative electrode state-of-charge identification results.\label{fig:SOC_0_05C}](SOC_0_05C_identification.png)
+
+## HPPC Identification
+Modfiy the `User Input` section in `Examples/Parameter_Identification_Routines/DFN_pso_HPPC.m`:
+```MATLAB
+%% User Input
+% Load parameters, including the stoichiometric parameter identified from C/20 discharge data
+% load('identified_parameters_0_05C.mat','param')
+
+% Enter mat file name where your PSO results will be stored
+file_name = 'pso_HPPC';
+ 
+% Enter names of parameters to identify (make sure names match the
+% parameter names in "param" structure containing nominal parameters)
[…]
