+++ paper/paper.md
@@ -61,9 +61,6 @@ In contrast, several open-source DFN model simulation tools have been released s
 
 # Examples 
 Visit COBRAPRO’s [Github](https://github.com/COBRAPROsimulator/COBRAPRO) to view all example codes:
-- **Parameter identification routine:** Utilizes PSO to optimize parameters using experimental current-voltage data
-- **DFN model implementation:** Finite volume method (FVM) discretization of the PDEs to form a DAE and SUNDIALS IDA solver to solve the DAE system
-
 
 - `Examples/Cycling`: contains battery cycling example codes
 - `Examples/Parameter_Identification_Routines`: contains parameter identification example codes
@@ -74,7 +71,7 @@ Here, we provide an overview of the parameter identification example codes in th
 Two example codes are provided that demonstrate a two-step parameter identification process. In step 1, the stoichiometric parameters are identified in `DFN_pso_0_05C.m` using the experimentally obtained C/20 discharge profile. In step 2, the electrolyte transport and kinetic parameters are identified in `DFN_pso_HPPC.m` using the experimentally obtained HPPC profile. 
 
 ## C/20 Discharge Identification
-`DFN_pso_0_05C.m`, located in the `Examples/Parameter_Identification_Routines` directory, is structured to require user input modifications only in the `User Input` section. This section includes definitions for parameter names, their respective upper and lower bounds, and experimental data:
+In `DFN_pso_0_05C.m`, only the `User Input` section needs to be modified. This section includes definitions for parameter names, their respective upper and lower bounds, and experimental data:
 ```MATLAB
 %% User Input  
 % Load nominal parameters 
@@ -122,7 +119,7 @@ SOC_init = 1;
 ```
 Note that `Parameters_LG_INR21700_M50.m` function defines the nominal parameters of your cell and the simulation settings, e.g., discretization method, DAE initialization method, current type, etc. 
 
-Once the user input has been defined, run the code to start the PSO. Once the PSO is finished, the code prints the identified parameter values and objective function values to the command window:
+Once the user input has been defined, run the `DFN_pso_0_05C.m` code to start the PSO. Once the PSO is finished, the code prints the identified parameter values and objective function values to the command window:
 ```
 Displaying identified values...
 ------------------------
@@ -150,7 +147,8 @@ J_SOCp =0.030231 [%]
 J_SOCn =0.019037 [%]
 J_tot =0.003833 [-]
 ```
-to the Command Window and plots the identified results as shown in \autoref{fig:V_0_05C} and \autoref{fig:SOC_0_05C}. 
+Model is simulated using the identified parameters and plotted with the experimental data as shown in \autoref{fig:V_0_05C} and \autoref{fig:SOC_0_05C}. 
+
 To view the C/20 identification results shown here, run 
 `Examples/Parameter_Identification_Results/DFN_pso_0_05C_identification`.
 
@@ -159,7 +157,7 @@ To view the C/20 identification results shown here, run
 ![C/20 discharge positive and negative electrode state-of-charge identification results.\label{fig:SOC_0_05C}](SOC_0_05C_identification.png){ width=60% }
 
 ## HPPC Identification
-Similar to `DFN_pso_HPPC.m`, modify the `User Input` section in `DFN_pso_HPPC.m`, located in the `Examples/Parameter_Identification_Routines` directory.
+Modify the `User Input` section in `DFN_pso_HPPC.m` to define the parameter names, parameter upper and lower bounds, and experimental data:
 ```MATLAB
 %% User Input
 % Load nominal parameters and stoichiometric parameter identified
@@ -224,7 +222,49 @@ V = V_data;
 % Enter initial SOC of your HPPC data
[…]
