+++ paper/paper.bib
@@ -150,6 +150,18 @@
 	doi = {10.4271/2023-01-5047},
 }
 
+@article{pozzato_data_2022,
+  title = {Lithium-Ion Battery Aging Dataset Based on Electric Vehicle Real-Driving Profiles},
+  author = {Pozzato, Gabriele and Allam, Anirudh and Onori, Simona},
+  year = {2022},
+  month = apr,
+  journal = {Data in Brief},
+  volume = {41},
+  pages = {107995},
+  issn = {23523409},
+  doi = {10.1016/j.dib.2022.107995},
+}
+
 @article{berliner_methods_2021,
   title = {Methods{\textemdash}{{PETLION}}: {{Open-Source Software}} for {{Millisecond-Scale Porous Electrode Theory-Based Lithium-Ion Battery Simulations}}},
   shorttitle = {Methods{\textemdash}{{PETLION}}},
