+++ paper/paper.md
@@ -45,13 +45,13 @@ In contrast, several open-source DFN model simulation tools have been released s
 
 ## Challenge 3. Unknown model parameters
 - **Issue:** As highlighted earlier, battery parameters are frequently unknown, and even if obtained through experimental characterization, parameter calibration is essential to accurately model the battery.
-- **Solution:** A co-simulation parameter optimization framework is implemented in COBRAPRO. The particle swarm optimization (PSO), a gradient-free population-based algorithm, is employed due to its suitability for nonlinear models like the DFN model. The PSO aims to optimize parameters by minimizing the cost function defined as the error between the experimental and simulated voltage and state-of-charge curves. COBRAPRO employs PSO using MATLAB’s Parallel Computing Toolbox to accelerate PSO convergence through multicore processing.
+- **Solution:** A co-simulation parameter optimization framework is implemented in COBRAPRO. The particle swarm optimization (PSO), a gradient-free population-based algorithm, is employed due to its suitability for nonlinear models like the DFN model. The PSO aims to optimize parameters by minimizing the objective function defined as the error between the experimental and simulated voltage and state of charge (SOC) curves. COBRAPRO employs PSO using MATLAB’s Parallel Computing Toolbox to accelerate PSO convergence through multicore processing.
 
 # Core Capabilities
 - **Parameter identification routine:** PSO optimizes parameters using experimental current-voltage data
 - **DFN model implementation:** PDEs discretized with finite volume method (FVM) and DAE system solved with SUNDIALS IDA 
 - **Solid particle radial discretization options:**
-  - FVM (3rd order Hermite interpolation utilized to account for sharp concentration gradients near the particle surface when calculating surface concentration [@xu_comparative_2023])
+  - FVM and 3rd order Hermite interpolation utilized to calculate particle surface concentration to account for sharp concentration gradients near the particle surface [@xu_comparative_2023] 
   - Finite difference method (FDM)
 - **DAE initialization options:**
   - Single-step approach [@lawder_extending_2015]
@@ -80,23 +80,31 @@ Visit COBRAPRO’s [Github](https://github.com/COBRAPROsimulator/COBRAPRO) to vi
   - `DFN_LSA_CC.m`: LSA on CC profile
   - `DFN_LSA_HPPC.m`: LSA on HPPC profile
   
-Here, we provide an overview of parameter identification examples located in the `Examples/Parameter_Identification_Routines` directory.
-Two example codes are provided to demonstrate a two-step parameter identification process using experimental data obtained from LG INR21700-M50T cells [@pozzato_data_2022]. In step 1, `DFN_pso_0_05C.m` identifies the stoichiometric parameters using C/20 discharge data. In step 2, `DFN_pso_HPPC.m` identifies the electrolyte transport and kinetic parameters using HPPC data. The remaining parameters are assumed to be known and obtained from LG INR21700-M50 cells [@chen_development_2020]. 
+Here, we provide a brief summary of the parameter identification examples located in the `Examples/Parameter_Identification_Routines` directory.
[…]
