+++ paper/paper.md
@@ -30,8 +30,7 @@ COBRAPRO (**Co**-simulation **B**atte**r**y Modeling for **A**ccelerated **P**a*
 
 Currently, DFN modeling tools lack integrated parameter identification routines, despite their need to accurately predict the dynamical behavior of batteries [@sulzer_python_2021] [@torchio_lionsimba_2016] [@smith_multiphase_2017] [@berliner_methods_2021]. 
 
-COMSOL Multiphysics&copy; [@comsol] is a commercially available finite element modeling software commonly used for DFN simulations. Although COMSOL does not include a built-in parameter identification feature, the *LiveLink&trade; for MATLAB&copy;* enables the creation of a co-simulation parameter optimization routine. *LiveLink&trade; for MATLAB&copy;*
-establishes communication between the two software platforms, allowing users to leverage the versatile suite of optimizers in MATLAB while running COMSOL to generate the model output [@pozzato_general_2023]. However, the expensive licensing fee and proprietary nature of COMSOL create barriers to public access, limiting collaboration and code reproducibility.
+COMSOL Multiphysics&copy; [@comsol] is a commercially available finite element modeling software commonly used for DFN simulations. Although COMSOL does not include a built-in parameter identification feature, the *LiveLink&trade; for MATLAB&copy;* enables the creation of a co-simulation parameter optimization routine by establishing communication between the two software platforms. This allows users to leverage the versatile suite of optimizers in MATLAB while running COMSOL to generate the model output [@pozzato_general_2023]. However, the expensive licensing fee and proprietary nature of COMSOL create barriers to public access, limiting collaboration and code reproducibility.
 
 In contrast, several open-source DFN model simulation tools have been released such as PyBaMM [@sulzer_python_2021], LIONSIMBA [@torchio_lionsimba_2016], PETLION [@berliner_methods_2021], DEARLIBS [@lee_robust_2021], fastDFN[@fastDFN], and MPET [@smith_multiphase_2017]. Among these packages, DEARLIBS is the only software with the capability to perform closed-loop parameter identification using experimental data. Other packages rely on literature-derived parameter values to generate simulation results. Taking inspiration from DEARLIBS, COBRAPRO aims to address three primary challenges in the DFN model:
 
