### 2024-04-08 #6540 @COBRAPROsimulator
Thank you @kyleniemeyer! Here is a list of potential reviewers who already agreed to review for JOSS:

- leeberty84
- yuefan98
- BradyPlanden
- DEARLIBS
- medullar
- yaomz16
- tomtranter
- EricaEgg

Here is a list of potential reviewers who are not on the JOSS reviewer list:
- petermattia
- martinjrobins
- robertwtimms
