### 2024-04-08 #6540 @COBRAPROsimulator
@kyleniemeyer, could you clarify if you mean that there is no JOSS editor suitable for COBRAPRO? Or do you mean that there is an appropriate JOSS editor for our code, but that they are not available at this moment (we noticed Rachel Kurchin could be a good match for our paper)?

If the latter, would you be able to give us an estimated time of when our paper would be assigned an editor?

Thank you,
Sara
