+++ paper/paper.bib
@@ -176,6 +176,26 @@
   doi = {10.1149/1945-7111/ac201c},
 }
 
+@article{ha_cobrapro_2024},
+  title = {COBRAPRO: An Open-source Software for the Dolye-Fuller-Newman Model with Co-simulation Parameter Optimization Framework},
+  author = {Ha, Sara and Onori, Simona},
+  journal = {Journal of The Electrochemical Society},
+  year = {2024},
+}
+
+@article{arunachalam_full_2019,
+  title = {Full {{Homogenized Macroscale Model}} and {{Pseudo-2-Dimensional Model}} for {{Lithium-Ion Battery Dynamics}}: {{Comparative Analysis}}, {{Experimental Verification}} and {{Sensitivity Analysis}}},
+  shorttitle = {Full {{Homogenized Macroscale Model}} and {{Pseudo-2-Dimensional Model}} for {{Lithium-Ion Battery Dynamics}}},
+  author = {Arunachalam, Harikesh and Onori, Simona},
+  year = {2019},
+  journal = {Journal of The Electrochemical Society},
+  volume = {166},
+  number = {8},
+  pages = {A1380-A1392},
+  issn = {0013-4651, 1945-7111},
+  doi = {10.1149/2.0051908jes},
+}
+
 @article{lee_robust_2021,
   title = {A {{Robust}} and {{Sleek Electrochemical Battery Model Implementation}}: {{A MATLAB}}{\textregistered} {{Framework}}},
   shorttitle = {A {{Robust}} and {{Sleek Electrochemical Battery Model Implementation}}},
@@ -247,3 +267,4 @@
   year = {2016},
   url = {https://github.com/scott-moura/fastDFN},
 }
+
