+++ paper/paper.bib
@@ -176,7 +176,7 @@
   doi = {10.1149/1945-7111/ac201c},
 }
 
-@article{ha_cobrapro_2024},
+@article{ha_cobrapro_2024,
   title = {COBRAPRO: An Open-source Software for the Dolye-Fuller-Newman Model with Co-simulation Parameter Optimization Framework},
   author = {Ha, Sara and Onori, Simona},
   journal = {Journal of The Electrochemical Society},
