+++ Examples/Local_Sensitivity_Analysis/DFN_LSA_Corr_CC.m
@@ -1,0 +1,268 @@
+% -------------------------------------------------------------------------
+% DFN_LSA_Corr_CC.m
+% -------------------------------------------------------------------------
+% DFN_LSA_Corr_CC performs local sensitivity analysis (LSA) for a CC profile
+% by perturbing each parameter one at a time. Also, correlation analysis is conducted
+% to quantify parameter correlation. Script will output identifiable
+% parameters using just LSA and using LSA correlation analysis. 
+% The sensitive/uncorrelated parameters can be identified using HPPC
+% experimental data in the "DFN_pso_CC.m" script.
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+% COBRAPRO: Co-simulation Battery Modeling for Accelerated Parameter Optimization
+
+% Copyright (c) 2024 CO-simulation BatteRy modeling for Accelerated PaRameter Optimization (COBRAPRO)
+% COBRAPRO is freely distributed software under the MIT License 
+% v1.0.0.: Released 02/2024 
+
+% Written by Sara Ha (sungyeon.sara.ha@stanford.edu)
+% PI: Prof. Simona Onori (sonori@stanford.edu)
+% Stanford Energy Control Group, Stanford University
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+ 
+%% User Input  
+clc;clear;close all
+
+%--------------------------------------------------------------------------
+% Load nominal/reference parameters
+%--------------------------------------------------------------------------
+param = Parameters_LG_INR21700_M50;
+
+%--------------------------------------------------------------------------
+% Enter names of parameters to conduct LSA (make sure names match the
+% parameter names in "param" structure containing nominal parameters)
+%--------------------------------------------------------------------------
+param_LSA_CC = {'ep' 'en' 'Rpp' 'Rpn'};
+
+% Type in latex version of parameter names for nice plots
+theta_names = {'$\varepsilon_p$' '$\varepsilon_n$' '$R_p$' '$R_n$'};
+
+%--------------------------------------------------------------------------
+% Perturbation coefficient for LSA [-]
+%--------------------------------------------------------------------------
+pct=0.05;
+
+%--------------------------------------------------------------------------
+% Define sensitivity index threshold
+%--------------------------------------------------------------------------
+% Parameters with sensitivity index higher than beta_LSA will be considered
+% the "identifiable" parameters for the HPPC profile
+%--------------------------------------------------------------------------
+beta_LSA = 0.5;
+
+%--------------------------------------------------------------------------
+% Define correlation coefficient threshold
+%--------------------------------------------------------------------------
+% Correlation coefficient threshold used to determine the uncorrelated parameter  
+% vector, which will be considered the "identifiable" parameters for the HPPC profile
+%--------------------------------------------------------------------------
[…]
