+++ Examples/Parameter_Identification_Routines/DFN_pso_HPPC.m
@@ -44,7 +44,7 @@ file_name = 'pso_HPPC';
 % parameter names in "param" structure containing nominal parameters)
 %--------------------------------------------------------------------------
 % You can also load identifiable parameter results from previous 
-% identifiability analysis conducted using "Examples\DFN_LSA_Corr_HPPC.m"
+% identifiability analysis conducted using "Examples/DFN_LSA_Corr_HPPC.m"
 %--------------------------------------------------------------------------
 param_HPPC = {'kp', 'kn', 'Dsp', 'Dsn', 'Kappa', 'De', 't1_constant', 'c0'};
 
