+++ Examples/Local_Sensitivity_Analysis/DFN_LSA_sim_HPPC.m
@@ -1,163 +1,0 @@
-function [t_nom,V_nom,SOCp_nom,SOCn_nom,...
-    t_l_all,V_l_all,SOCp_l_all,SOCn_l_all,...
-    t_u_all,V_u_all,SOCp_u_all,SOCn_u_all,...
-    S_norm_V, S_norm_SOCp, S_norm_SOCn, S_norm_V_SOC]...
-    = DFN_LSA_sim_HPPC(param,param_LSA_HPPC,theta_nom,pct)
-%   DFN_LSA_sim_HPPC conducts local sensitivity analysis (LSA) by locally 
-%   perturbing the parameters of interest around a nominal value for HPPC simulations. 
-%
-%   Inputs:
-%       param: parameter structure
-%       param_LSA_HPPC: cell consisting of parameter names to conduct LSA on
-%       theta_nom: vector consisting of nominal values of the parameters to conduct LSA on
-%       pct: perturbation coefficient [-]
-%
-%   Output:
-%       t_nom: time vector simulated using nominal parameters [s]
-%       V_nom: voltage vector simulated using nominal parameters [V]
-%       SOCp_nom: Positive electrode SOC vector simulated using nominal parameters [-]
-%       SOCn_nom: Negative electrode SOC vector simulated using nominal parameters [-]
-%       t_l_all: cell array where each cell contains the time vector simulated using the lower perturbed parameter [s]
-%       V_l_all: cell array where each cell contains the voltage vector simulated using the lower perturbed parameter [V]
-%       SOCp_l_all: cell array where each cell contains the positive electrode SOC vector simulated using the lower perturbed parameter [-]
-%       SOCn_l_all: cell array where each cell contains the negative electrode SOC vector simulated using the lower perturbed parameter [-]
-%       t_u_all: cell array where each cell contains the time vector simulated using the upper perturbed parameter [s]
-%       V_u_all: cell array where each cell contains the voltage vector simulated using the upper perturbed parameter [V]
-%       SOCp_u_all: cell array where each cell contains the positive electrode SOC vector simulated using the upper perturbed parameter [-]
-%       SOCn_u_all: cell array where each cell contains the negative electrode SOC vector simulated using the upper perturbed parameter [-]
-%       S_norm_V: normalized voltage sensitivity vector [-]
-%       S_norm_SOCp: normalized positive electrode SOC sensitivity vector [-]
-%       S_norm_SOCn: normalized negative electrode SOC sensitivity vector [-]
-%       S_norm_V_SOC: normalized total sensitivity vector (including voltage, positive electrode SOC, and negative electrode SOC sensitivities) [-]
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% COBRAPRO: Co-simulation Battery Modeling for Accelerated Parameter Optimization
-
-% Copyright (c) 2024 CO-simulation BatteRy modeling for Accelerated PaRameter Optimization (COBRAPRO)
-% COBRAPRO is freely distributed under the MIT License 
-% v1.0.0.: Released March, 2024 
-
-% Main contributors: 
-% Sara Ha (sungyeon.sara.ha@stanford.edu)
-% Simona Onori (sonori@stanford.edu)
-% Stanford Energy Control Group, Stanford University
[…]
