+++ Examples/Local_Sensitivity_Analysis/DFN_LSA_HPPC.m
@@ -1,206 +1,0 @@
-% -------------------------------------------------------------------------
-% DFN_LSA_HPPC.m
-% -------------------------------------------------------------------------
-%   DFN_LSA_HPPC performs local sensitivity analysis (LSA) for HPPC by perturbing 
-%   each parameter one-at-a-time. This script can be used to determine sensitive
-%   parameters for the HPPC profile. Once the sensitive parameters are determined, they can be 
-%   identified using HPPC experimental data using "DFN_pso_HPPC.m" script.
-%   This script takes longer to run than DFN_LSA_CC.m since we are simulating HPPC profile.
-
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% COBRAPRO: Co-simulation Battery Modeling for Accelerated Parameter Optimization
-
-% Copyright (c) 2024 CO-simulation BatteRy modeling for Accelerated PaRameter Optimization (COBRAPRO)
-% COBRAPRO is freely distributed under the MIT License 
-% v1.0.0.: Released March, 2024 
-
-% Main contributors: 
-% Sara Ha (sungyeon.sara.ha@stanford.edu)
-% Simona Onori (sonori@stanford.edu)
-% Stanford Energy Control Group, Stanford University
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
- 
-%% User Input  
-clc;clear;close all
-
-%--------------------------------------------------------------------------
-% Load nominal/reference parameters
-%--------------------------------------------------------------------------
-param = Parameters_LG_INR21700_M50;
-param.Rc = 0;
-
-%--------------------------------------------------------------------------
-% Enter names of parameters to conduct LSA (make sure names match the
-% parameter names in "param" structure containing nominal parameters)
-%--------------------------------------------------------------------------
-param_LSA_HPPC = {'Dsp' 'Dsn' 't1_constant' 'kp' 'kn' 'c0' 'De' 'Kappa'};
-
-% Type in latex version of parameter names for nice plots
-theta_names = {'$D_{s,p}$' '$D_{s,n}$' '$t_+$' '$k_p$' '$k_n$' '$c_0$' '$D_e$' '$\kappa$'};
-
-%--------------------------------------------------------------------------
-% Perturbation coefficient for LSA [-]
-%--------------------------------------------------------------------------
-pct=0.05;
-
-%--------------------------------------------------------------------------
-% Load HPPC profile input current vector [A] and simulation time [s] 
-%--------------------------------------------------------------------------
-% time_vec: Should be a vector where each entry defines the time period for each CC segment [s] (Px1)
-% curr_vec: Should be a vector where each entry defines the current for that CC segment [A] (Px1) (negative current: discharging)
-% -> where P is the number of CC segments for your HPPC experiment
-%--------------------------------------------------------------------------
-% For example, since HPPC starts with a 1C discharge for 360 s followed by
[…]
