+++ Examples/Local_Sensitivity_Analysis/DFN_LSA_Corr_HPPC.m
@@ -1,287 +1,0 @@
-% -------------------------------------------------------------------------
-% DFN_LSA_Corr_HPPC.m
-% -------------------------------------------------------------------------
-% DFN_LSA_Corr_HPPC performs local sensitivity analysis (LSA) for the HPPC profile
-% by perturbing each parameter one at a time. Also, correlation analysis is conducted
-% to quantify parameter correlation. Script will output identifiable
-% parameters using just LSA and using LSA correlation analysis. 
-% The sensitive/uncorrelated parameters can be identified using HPPC
-% experimental data in the "DFN_pso_HPPC.m" script.
-%
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-% COBRAPRO: Co-simulation Battery Modeling for Accelerated Parameter Optimization
-
-% Copyright (c) 2024 CO-simulation BatteRy modeling for Accelerated PaRameter Optimization (COBRAPRO)
-% COBRAPRO is freely distributed software under the MIT License 
-% v1.0.0.: Released 02/2024 
-
-% Written by Sara Ha (sungyeon.sara.ha@stanford.edu)
-% PI: Prof. Simona Onori (sonori@stanford.edu)
-% Stanford Energy Control Group, Stanford University
-%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
- 
-%% User Input  
-clc;clear;close all
-
-%--------------------------------------------------------------------------
-% Option 1: Enter identified parameters
-%--------------------------------------------------------------------------
-load('identified_parameters_0_05C.mat','param')
-param.upperCutoffVoltage = 4.3; 
-%--------------------------------------------------------------------------
-% Option 2: If you don't have identified parameters, then load parameters from literature [Chen 2020]
-%--------------------------------------------------------------------------
-% param = Parameters_LG_INR21700_M50;
-
-%--------------------------------------------------------------------------
-% Option 1: Enter names of parameters to conduct LSA (make sure names match the
-% parameter names in "param" structure containing nominal parameters)
-%--------------------------------------------------------------------------
-% param_LSA_HPPC = {'Dsp' 'Dsn' 't1_constant' 'kp' 'kn' 'c0' 'De' 'Kappa'};
-%--------------------------------------------------------------------------
-% Option 2: Load identifiable parameters from identifiability analysis
-% conducted in "Examples/Local_Sensitivity_Analysis/DFN_LSA_Corr_HPPC.m"
-%--------------------------------------------------------------------------
-load('DFN_identification_results/HPPC_identifiable_params.mat','LSA_identifiable','corr_identifiable')
-% 1. Load results from LSA analysis 
-param_LSA_HPPC = LSA_identifiable;
-% 2. Load results from LSA + correlation analysis 
-% param_LSA_HPPC = corr_identifiable;
-
-% Type in latex version of param_LSA_HPPC for nice plots
-theta_names = {'$D_{s,p}$' '$k_n$' '$c_0$' '$D_{s,n}$'};
-
-%--------------------------------------------------------------------------
-% Perturbation coefficient for LSA [-]
[…]
