+++ paper/paper.md
@@ -64,11 +64,13 @@ In contrast, several open-source DFN model simulation tools have emerged, such a
 
 # Example: Case Study on LG 21700-M50T Cells
 
-As a demonstration of COBRAPRO, we conduct a case study aimed at parameterizing a fresh LG 21700-M50T cell using C/20 capacity test, HPPC, and driving cycle data [@pozzato_data_2022]. In this example, we break down the identification problem by systematically grouping parameters for each identification step, as shown in \autoref{fig:flowchart}. This multi-step approach aims to improve the identifiability of parameters instead of identifying all the unknown parameters simulataneously [@arunachalam_full_2019]. 
+As a demonstration of COBRAPRO, we conduct a case study aimed at parameterizing a fresh LG 21700-M50T cell using the C/20 capacity test, HPPC, and driving cycle data [@pozzato_data_2022]. In this example, we break down the identification problem by systematically grouping parameters for each identification step, as shown in \autoref{fig:flowchart}. This multi-step approach aims to improve the identifiability of parameters instead of identifying all the unknown parameters simulataneously [@arunachalam_full_2019]. 
 
-![Case study: Parameter identification procedure on LG 21700-M50T cells.\label{fig:flowchart}](example_flowchart_v3.pdf){ width=100% }
+![Case study: Parameter identification procedure on LG 21700-M50T cells.\label{fig:flowchart}](example_flowchart.pdf){ width=100% }
 
 First, the geometric parameters and open-circuit potential functions are extracted from measurements conducted in cell tear-down and half-cell experiments on LG 21700-M50 cells, as reported by [@chen_development_2020]. Next, the C/20 capacity test data is used to identify the stoichiometric parameters, as shown in the example code `DFN_pso_0_05C.m`. Then, we conduct a parameter identifiability study comprising of LSA and correlation analysis to pinpoint parameters with high sensitivity to voltage and SOC while maintaining low correlation with other parameters. Further details on parameter identifiability can be found in [@ha_cobrapro_2024]. Next, we identify the electrolyte transport and kinetic parameters using HPPC data in the example code `DFN_pso_HPPC.m`. Finally, validation of the identified parameters is carried out on UDDS data using the code `DFN_UDDS_validation.m`. The `DFN_pso_0_05C.m` and `DFN_pso_HPPC.m` files are located in the `Examples/Parameter_Identification_Routines` directory and `DFN_UDDS_validation.m` is located in `Examples/Parameter_Identification_Results`.
+
+## C/20 Capacity Test Identification
 
 Here, we provide a quick overview of `DFN_pso_0_05C.m`. In `DFN_pso_0_05C.m`, the `User Input` section is used to define the parameter names, their respective upper and lower bounds, experimental data, PSO settings, etc. Note that `DFN_pso_HPPC.m` and `DFN_UDDS_validation.m` also consist of `User Input` at the beginning of the code, which is the only part that needs to be modified by the user. A preview of the `User Input` section is provided here.
 
@@ -149,16 +151,13 @@ J_SOCp =0.030231 [%]
 J_SOCn =0.019037 [%]
 J_tot =0.003833 [-]
 ```
-The code also outputs the plots of the simulation results generated from the identified parameters and the experimental data, as shown in \autoref{fig:flowchart}. <!--T \autoref{fig:V_0_05C} and \autoref{fig:SOC_0_05C}. -->
-<!--T
[…]
