+++ paper/paper.bib
@@ -179,7 +179,7 @@
 @article{ha_cobrapro_2024,
   title = {COBRAPRO: An Open-source Software for the Dolye-Fuller-Newman Model with Co-simulation Parameter Optimization Framework},
   author = {Ha, Sara and Onori, Simona},
-  journal = {Journal of The Electrochemical Society},
+  journal = {Journal of The Electrochemical Society, under review},
   year = {2024},
 }
 
