### 2024-04-26 #6540 @mbarzegary
Hi @BradyPlanden, @yuefan98, @DEARLIBS, @yaomz16, @EricaEgg 👋
Considering your expertise, would any of you be willing to review this submission for JOSS? As you may know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06540/joss.06540/10.21105.joss.06540.pdf
Software: https://github.com/COBRAPROsimulator/COBRAPRO
