### 2024-04-26 #6540 @yuefan98
Hi Mojtaba,

I would be happy to help with this. Is there a specific deadline for this
review.

Best,
Yuefan Ji

On Fri, Apr 26, 2024 at 9:16 AM Mojtaba Barzegari ***@***.***>
wrote:

> Hi @BradyPlanden
> <https://urldefense.com/v3/__https://github.com/BradyPlanden__;!!K-Hz7m0Vt54!gSLIWm0L-iXK9ojzwD7QGfsBHcS27mBcZJhLfCessOm_av95yTuiwxdwSJpwvIPj5NoVZok-pOtfJypIGIH1hgE$>,
> @yuefan98
> <https://urldefense.com/v3/__https://github.com/yuefan98__;!!K-Hz7m0Vt54!gSLIWm0L-iXK9ojzwD7QGfsBHcS27mBcZJhLfCessOm_av95yTuiwxdwSJpwvIPj5NoVZok-pOtfJypIh5NTIec$>,
> @DEARLIBS
> <https://urldefense.com/v3/__https://github.com/DEARLIBS__;!!K-Hz7m0Vt54!gSLIWm0L-iXK9ojzwD7QGfsBHcS27mBcZJhLfCessOm_av95yTuiwxdwSJpwvIPj5NoVZok-pOtfJypItUSZlhI$>,
> @yaomz16
> <https://urldefense.com/v3/__https://github.com/yaomz16__;!!K-Hz7m0Vt54!gSLIWm0L-iXK9ojzwD7QGfsBHcS27mBcZJhLfCessOm_av95yTuiwxdwSJpwvIPj5NoVZok-pOtfJypInY-1dSU$>,
> @EricaEgg
> <https://urldefense.com/v3/__https://github.com/EricaEgg__;!!K-Hz7m0Vt54!gSLIWm0L-iXK9ojzwD7QGfsBHcS27mBcZJhLfCessOm_av95yTuiwxdwSJpwvIPj5NoVZok-pOtfJypIg2nBJuQ$>
> 👋
> Considering your expertise, would any of you be willing to review this
> submission for JOSS? As you may know, we carry out our checklist-driven
> reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
> <https://urldefense.com/v3/__https://joss.readthedocs.io/en/latest/review_criteria.html__;!!K-Hz7m0Vt54!gSLIWm0L-iXK9ojzwD7QGfsBHcS27mBcZJhLfCessOm_av95yTuiwxdwSJpwvIPj5NoVZok-pOtfJypIW4T0LBc$>
>
> Paper:
> https://raw.githubusercontent.com/openjournals/joss-papers/joss.06540/joss.06540/10.21105.joss.06540.pdf
> <https://urldefense.com/v3/__https://raw.githubusercontent.com/openjournals/joss-papers/joss.06540/joss.06540/10.21105.joss.06540.pdf__;!!K-Hz7m0Vt54!gSLIWm0L-iXK9ojzwD7QGfsBHcS27mBcZJhLfCessOm_av95yTuiwxdwSJpwvIPj5NoVZok-pOtfJypItq2Xry4$>
> Software: https://github.com/COBRAPROsimulator/COBRAPRO
> <https://urldefense.com/v3/__https://github.com/COBRAPROsimulator/COBRAPRO__;!!K-Hz7m0Vt54!gSLIWm0L-iXK9ojzwD7QGfsBHcS27mBcZJhLfCessOm_av95yTuiwxdwSJpwvIPj5NoVZok-pOtfJypIVoxsMKs$>
>
