### 2024-04-26 #6540 @yaomz16
Hi Mojtaba,

Sorry I cannot review this paper for now, as I am extremely busy these
days.

Thanks,

Archie (yaomz16)

On Fri, Apr 26, 2024 at 11:16 AM Mojtaba Barzegari ***@***.***>
wrote:

> Hi @BradyPlanden <https://github.com/BradyPlanden>, @yuefan98
> <https://github.com/yuefan98>, @DEARLIBS <https://github.com/DEARLIBS>,
> @yaomz16 <https://github.com/yaomz16>, @EricaEgg
> <https://github.com/EricaEgg> 👋
> Considering your expertise, would any of you be willing to review this
> submission for JOSS? As you may know, we carry out our checklist-driven
> reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Paper:
> https://raw.githubusercontent.com/openjournals/joss-papers/joss.06540/joss.06540/10.21105.joss.06540.pdf
> Software: https://github.com/COBRAPROsimulator/COBRAPRO
>
