### 2024-04-30 #6540 @mbarzegary
Hi @TomTranter 👋
Do you have time for reviewing another battery modeling work for JOSS?

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06540/joss.06540/10.21105.joss.06540.pdf
Software: https://github.com/COBRAPROsimulator/COBRAPRO
