### 2024-05-06 #6540 @mbarzegary
Hi @ma-sadeghi 👋
Do you have time for reviewing another battery modeling work for JOSS?

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06540/joss.06540/10.21105.joss.06540.pdf
Software: https://github.com/COBRAPROsimulator/COBRAPRO
