### 2024-05-14 #6540 @mbarzegary
Hi @jepegit 👋
Since you have recently published your software, would you be willing to review a battery modeling work for JOSS?

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06540/joss.06540/10.21105.joss.06540.pdf
Software: https://github.com/COBRAPROsimulator/COBRAPRO
