### 2024-05-29 #6803 @editorialbot
[REVIEW]: COBRAPRO: A MATLAB toolbox for Physics-based Battery Modeling and Co-simulation Parameter Optimization
**Submitting author:** @COBRAPROsimulator (<a href="http://orcid.org/0009-0005-9878-3537">Sara Ha</a>)
**Repository:** https://github.com/COBRAPROsimulator/COBRAPRO
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
