### 2024-06-03 COBRAPROsimulator/COBRAPRO#1 @yuefan98
Unrecognized casadiMex with ARM version MATLAB 2023b or later
I followed the installation guide and got the output for a successful installation on ARM version MATLAB 2023b and 2024a. However, I encountered issues with casadiMex and was unable to run the code with these two versions of MATLAB.

The installation and operation work with intel version MATLAB 2022b running on the same Mac with an arm chip. This is likely an issue with the arm version of MATLAB, so it might be worth troubleshooting or at least mentioning it in the known issues section.
