### 2024-06-06 COBRAPROsimulator/COBRAPRO#1 @yuefan98
Yes, here is the screenshot of the Matlab versions that are available from the website for the latest release. MATLAB can now run natively on Apple silicon, which is based on ARM architecture, but that seems to cause a problem with casadiMex. I don't know why that is the case. So it might be worth troubleshooting or at least mentioning in the documentation. The code can run without any problem with MATLAB 2022a (developed for intel processor) with the same mac I have with Apple silicon processor.
