### 2024-06-06 COBRAPROsimulator/COBRAPRO#1 @COBRAPROsimulator
Thank you for clarifying @yuefan98. Really appreciate you pointing out this issue. I will add to the known issues in the README page.
