### 2024-06-06 #6803 @COBRAPROsimulator
@yuefan98, thank you. Here is my comment to 4:

- Could you please clarify what you mean by validating HPPC results by reducing it to two particles optimization? Do you mean that you only used 2 particles in the PSO? What was your the optimized objective function value (RMSE between simulated and experimental voltage, J_V) in this case? And approximately how long did your optimization take? Also, were you able to validate the identified HPPC values using the UDDS profile? As you may know, in parameter optimization, validation is typically carried on a different profile (e.g., driving cycle) than the one it was identified with.
- To answer your point on solid phase diffusion, I would like to add a general comment about the DFN model parameterization pipeline we proposed. The DFN model is only a simplified representation of a real battery and makes several assumptions such as perfectly spherical particles, dynamics only in the x- and r-directions, and homogenous electrode properties. In my PhD work, I have seen that when we take experimentally measured parameters from cell-tear down and directly input them into the model, the model cannot predict the behavior of the real cell. This motivates the need for parameter identification. However, we cannot and should not identify/optimize all the DFN parameters at once since the DFN model is overparameterized. We use a layered identification technique to improve parameter identifiability, where the geometrical parameters are taken from cell-tear down measurements, stoichiometric parameters extracted from C/20 data, and the remaining transport and kinetic parameters are identified with the HPPC profile. For the HPPC identification, it is important to note that prior to conducting HPPC PSO, we conduct practical identifiability analysis to determine the identifiable parameters that will be optimized using the HPPC data. The identifiability analysis is conducted through LSA and correlation analysis in the Examples/Local_Sensitivity_Analysis/DFN_LSA_Corr_HPPC.m file. The user can use this code as a tool to determine the "identifiable" parameters, which then can be optimized in the Examples/Parameter_Identification_Routines/DFN_pso_HPPC.m file. This is pipeline is summarized in Fig 1 of the paper.
- To summarize, the goal is not for the identified transport and kinetic parameters to match the actual/real values measured from cell-tear down, since the DFN model is just a simplified representation of a real battery, but rather to identify solid phase diffusion coefficient that will reduce the experimental and simulated voltage and state-of-charge RMSE. Furthermore, this is why the UDDS validation is carried out to check that the HPPC identified parameters work well on a new/different profile.
- Our ECS paper, developed to complement the COBRAPRO code, explains this DFN model parameterization in detail. Also, it is important to note that this code was developed to be a tool to carry out parameter identifiability and identification, but the actual parameters being identified may change depending on the user's experimental profile and cell chemistry.
- You mention a good point that the in the current code, the solid phase diffusion is not a function of SOC. We can add this feature in the future but will require some development since the radial discretization method will need to be reconstructed to allow SOC dependent Dsp and Dsn coefficients.

I hope I answered your question. Please let me know if you have any other questions.
