### 2024-06-07 #6803 @yuefan98
- Yes, I just ran your cycle_HPPC.m with 2 particles in PSO, and cycle_UDDS.m. By reducing it to two particles I can ran your code in around ~ 10 mins, but I think that's sufficient for me to validate your code. I think I am all good with that
- Thanks for the clarification. I agree that battery parameterization are complicated and all those parameters can coupled together. I think your pipeline is good and reasonable.
- I think everything is good from my side at this point.
