### 2024-07-02 #6803 @brosaplanella
> [!WARNING]
> This review is WIP, I keep publishing & editing the post at regular intervals to ensure it is saved. I will remove this warning once it is complete.

Here's my review. Overall, the package provides a novel solution to the parameterisation of physics-based battery models (in this case the DFN). The code is functional and the article well-written, and meets all the points in the checklist (though note I have some questions regarding the performance, see below). I provide a list of comments to address before publication, split into major and minor comments.

# Major comments
- One of the main claims in the paper is performance, and it is stated that the 1C discharge runs in less than a second (similar order of magnitude to PyBaMM). When running it on my laptop (it tic/toced the first 1C discharge in `cycle_CC.m` it took around 8 seconds (the PyBaMM time in my machine is similar to that reported by the authors). I understand that the discrepancy is probably due to pre/post-processing and printing to screen, but it would be good to have one example with minimal overhead which can demonstrate the performance claimed in the article.
- I agree with @yuefan98 that the documentation is a bit sparse. Even though for a research tool the current documentation might be enough, it should be expanded in the near future (I am also happy for this to take place after publication). I believe the main issue is that, even though the examples work straight out of the box, it would be extremely hard for a user to adapt the code for their own needs. I believe that publishing some MATLAB notebooks commenting on the steps would be very useful.
