### 2024-07-10 #6803 @mbarzegary
@COBRAPROsimulator can you please provide an update on the above, in terms of responding to the issues raised by @brosaplanella?
