### 2024-08-02 #6803 @BradyPlanden
Apologies for the delay, here is my review in full.

Overall, this paper is well written and makes a novel contribution to the subfield of parameterisation for electrochemical battery modelling. As the current work has all been completed through the @COBRAPROsimulator account, I have assumed that is the first author, as Ferran mentioned above, setting up a contributor list would be helpful and enable expansion in the future.

## Repository Comments

- While the authors provide an initial testing framework, this only covers conventional integration tests and lacks individual unit tests or providing users with a coverage metric.
- In it's current state, the repository appears to be more of a collection of scripts rather than a package/module/etc. This makes it difficult to determine the software architecture, potentially limiting the reach of this work. This repository provides a reference to a possible structure (as a MATLAB codebase): https://github.com/Battery-Intelligence-Lab/BatEst
- More functionality could be added to the particle swarm optimisation object (i.e. max iterations, time, etc.) with documentation. Specifically, here: https://github.com/COBRAPROsimulator/COBRAPRO/blob/157ab80ffda30aa7c51c39cc7a042162efbfbaf0/Examples/Parameter_Identification_Routines/DFN_pso_0_05C.m#L198

## Article Comments
- Presently, the comparison with alternative work in the field is aimed at pure forward modelling packages (PyBaMM, PetLion.jl, LionSimba) which do not aim to solve the inverse problem, a more extensive literature review  will find other implementations which are more suitable for comparison.
- A table showing the full set of parameters identified for the example would be helpful to summarise the results in one place for the reader.
- Similarly, a table of correlation metrics for the parameters in the LSA section of the example.
