### 2024-08-12 #6803 @mbarzegary
@COBRAPROsimulator can you please provide an update on the raised issues? We need to move forward with this submission.
