### 2024-08-12 #6803 @COBRAPROsimulator
@brosaplanella , thank you so much for your detailed review. I apologize for my delayed response. Here are my responses below:

Major comments:
- Comment 1. One of the main claims in the paper is performance, and it is stated that the 1C discharge runs in less than a second (similar order of magnitude to PyBaMM). When running it on my laptop (it tic/toced the first 1C discharge in cycle_CC.m it took around 8 seconds (the PyBaMM time in my machine is similar to that reported by the authors). I understand that the discrepancy is probably due to pre/post-processing and printing to screen, but it would be good to have one example with minimal overhead which can demonstrate the performance claimed in the article.
- Answer 1. Thank you for your comment. In the cycle_CC.m file, the "[output,param] = runModel" function calculates output, which is a structure that reports the time it look to solve the DFN model. The exact time is reported as "output.DAE_solve_time" in the "output" structure. This calculates the pure time it took to solve the DFN model (not including printing and code overhead).  I can modify the cycle_CC.m file so that it also prints the time to solve the DFN model. This should be less than one second for one cycle of 1C discharge for Np=Ns=Nn=Nrp=Nrn=10. We conducted a systematic computation study in our accepted JES paper which will be published soon.

- Comment 2: I agree with @yuefan98 that the documentation is a bit sparse. Even though for a research tool the current documentation might be enough, it should be expanded in the near future (I am also happy for this to take place after publication). I believe the main issue is that, even though the examples work straight out of the box, it would be extremely hard for a user to adapt the code for their own needs. I believe that publishing some MATLAB notebooks commenting on the steps would be very useful.
- Answer 2. Thank you for your comment. The author will add more documentation, like MATLAB notebooks, to improve the code adaptability and usability.

- Comment 3: Similarly, tests are quite sparse. At the moment the tests only check that CasADI runs and that one particular example works, but these need to be ran manually. I would recommend some continuous integration on Github (or similar) to test things work periodically, but I don't know if that is feasible with proprietary software like MATLAB. In any case, some additional tests should be written, at the very least one to check that the parameter fitting tools work as expected.
- Answer 3: Thank you for this comment. The author will add another test to ensure that the parameter optimization tool works properly.

- Comment 4: At the moment the list of CONTRIBUTORS is a bit opaque. All the commits have been pushed by COBRAPROsimulator, which does not tell who that is or if is one or multiple users. This could be an issue if other people join the project or the person behind COBRAPROsimulator changes institutions. I would recommend having a list of contributors in the README, where contributors are acknowledged. The [all-contributors](https://allcontributors.org/) tool is quite useful for this.
- Answer 4: Thank you for this comment. We will list the contributors in the README file.

Minor comments:

- Comment 5: The following comments refer to the installation instructions.
- Comment 5.1:The instructions require SUNDIALS 2.6.2 but the hyperlink points to 2.6.1. In addition, this version is quite old. Is there any particular reason to not use a more recent version?
- Answer 5.1: Thank you for pointing this out. The authors will fix the hyperlink so that SUNDIALS 2.6.2 is downloaded. In our accepted JES paper (hopefully published soon), we explain in detail why the old version of SUNDIALS is used. The robust single-step approach is used to determine consistent initial conditions [M. T. Lawder, V. Ramadesigan, B. Suthar, and V. R. Subramanian, “Extending explicit and linearly implicit ODE solvers for index-1 DAEs,” Computers & Chemical Engineering, vol. 82, pp. 283–292, Nov. 2015, doi: [10.1016/j.compchemeng.2015.07.002](https://doi.org/10.1016/j.compchemeng.2015.07.002).] in COBRAPRO. This method request the use of implicit ODE solver, which the Casadi version of SUNDIALS (what pybamm seems to use) cannot support. Casadi version of SUNDIALS IDA solver only allows semi-explicit ODE or DAE solvers, while the SUNDIALS 2.6.2 version allows for implicit ODE/DAE solvers.
- Comment 5.2: The instructions say CasADI should be installed, but they don't specify for which platform. Even though it is quite obvious, it would be helpful to specify it needs to be the MATLAB version.
- Answer 5.2: Yes sure, this can be added to the README file.
- Comment 5.3: The instructions assume MATLAB is installed, with some toolboxes. It should be stated explicitly at the beginning of the "Installation" section.
- Answer 5.3: The authors will edit the README file such that it explicitly says that MATLAB and relevant toolboxes should be installed sat the beginning of the “Installation” section.

- Comment 6: The following comments refer to the installation process in MATLAB:
- Comment 6.1: There is a step that asks "Install toolbox" that it is not reflected in the installation instructions.
- Answer 6.1: Thank you for your comment. This part will be added to the README file..
- Comment 6.2: It would be good to always have a default answer to most of the prompts (so just hitting enter gets you through), as the fact that some answers require a yes and some a no is a bit confusing. If that is already implemented, users should be made aware of what the default is.
- Answer 6.2: The authors will try to modify the installation file such that yes/no answers can be averted.

- Comment 7: The title of the correlation matrix plot (e.g. when running DFN_LSA_Corr_CC.m) does not render LaTeX.
- Answer 7: Thank you for catching that. We will fix that error.

- Comment 8: The DFN_pso_0_05C.m also took a long time to me (around 30 minutes) and some errors/warning came up. It would be good to clarify in the documentation that this example might take a while to run.
- Answer 8: Thank you for your comment. The authors will make it clear that the parameter optimization codes take a while to run. The errors/warnings that are printed to the screen are normal, as they notify the user why the simulation failed to work for the particular set of parameters in a PSO iteration.

- Comment 9: The following comments refer to the article:
- Comment 9.1: [l33] I would remove all the references here, as they are all cited again and discussed in more detail later on.
- Answer 9.1: The authors will remove the reference here.
- Comment 9.2: [l47] "Other packages resort to literature-derived parameters and lack the ability to predict battery behaviour". I think this statement is a bit of a stretch. All the aforementioned tools will provide models with the ability to predict real battery behaviour, as long as the user provides the right parameter values. COBRAPRO provides tools to estimate those parameters (which is no small feat), while the others don't, so I think this would be a more accurate statement.
- Answer 9.2: Thank you for pointing this out. The authors will modify the paper so that we highlight that COBRAPRO is a closed-loop code that allows for parameter identification, whereas other codes are open-loop which already require correct parameters.
- Comment 9.3: [l54]: "COBRAPRO leverages a fast solver". Which one? Give details and references.
- Answer 9.3: The authors will clarify that COBRAPRO uses the SUNDIALS IDA solver. We will also reference our JES paper which provides more information on the  SUNDIALS solver.
- Comment 9.4: [l64] About Challenge 2, I think it is a bit misleading. I suspect most of the other tools also look for consistent initial conditions (PyBaMM definitely does, and I suspect COMSOL does too; haven't played much with the other tools). If any of the other tools does not do that, it should be stated clearly. Otherwise, I think this challenge would be better merged with Challenge 1, as it all ties to dealing with the computational complexity.
- Answer 9.4: Unlike other DFN software, COBRAPRO implements the robust single-step approach [Lawder et al., 2015] which perturbs the algebraic variables into ODEs to prevent the use of nonlinear algebraic solvers and shown to have convergence issues. We have conducted a comparison study of the single-step method and the SUNDIALS IDACalc function, which uses a nonlinear algebraic solver, and showed that the single-step method implemented in COBRAPRO is more robust. This framework was first implemented in Maple in [Lawder et al., 2015] and implemented in MATLAB by DEARLIBS. We will provide more details on this in the JOSS paper.
- Comment 9.5: [l185, but there are other occurrences through the text] The subscripts denoting labels (e.g. n, p, e...) should be in Roman text (i.e. \mathrm{e}) in LaTeX.
- Answer 9.5.: Thank you for this comment. We will fix the paper accordingly.
- Comment 9.6: [Fig3&5] The plots showing the SOC are not very insightful. I imagine they are computed through Coulomb counting in both cases, and given that the experiment is current driven they need to be virtually identical. I believe a plot of the voltage error between experiment and simulation would be a more valuable output of the code.
- Answer 9.6: Thank you for this comment. COBRAPRO leverages the multi-objective cost function proposed by [A. Allam and S. Onori, “Online Capacity Estimation for Lithium-Ion Battery Cells via an Electrochemical Model-Based Adaptive Interconnected Observer,” IEEE Trans. Contr. Syst. Technol., vol. 29, no. 4, pp. 1636–1651, Jul. 2021, doi: [10.1109/TCST.2020.3017566](https://doi.org/10.1109/TCST.2020.3017566).], which showed increased sensitivity of parameters by including SOCp and SOCn in the voltage cost function. This is why COBRAPRO outputs the voltage and also the SOC of the electrodes.
- Comment 9.7: [l246] Given that this list is bound to change as the code evolves, I think it would be better to not include it in the article and, instead, point the users to the README where an up-to-date list will be available.
- Answer 9.7: Thank you for this valuable comment. The authors will remove this from the paper.
