### 2024-08-12 #6803 @COBRAPROsimulator
@BradyPlanden, thank you so much for your review. Here are our answers to your comments:

Repository Comments

- Comment 1: While the authors provide an initial testing framework, this only covers conventional integration tests and lacks individual unit tests or providing users with a coverage metric.
- Answer 1: Thank you for this comment. Could the reviewer please explain more in detail what additional testing codes should be developed to improve the code? Currently, the authors included a test to ensure that the DFN model is simulated correctly by including a comparison with COMSOL results. If this test is successful, it implies that all the sub function files are working well as well. Following @brosaplanella comment, the author will include an additional test to ensure that the parameter optimization is working properly.

- Comment 2: In its current state, the repository appears to be more of a collection of scripts rather than a package/module/etc. This makes it difficult to determine the software architecture, potentially limiting the reach of this work. This repository provides a reference to a possible structure (as a MATLAB codebase): https://github.com/Battery-Intelligence-Lab/BatEst
- Answer 2: Thank you for this comment. The author will improve the organization of COBRAPRO such that the software architecture is easier to comprehend. For example, the author can create subfolders in the DFN_model folder to provide more granular information on the code structure.

- Comment 3: More functionality could be added to the particle swarm optimisation object (i.e. max iterations, time, etc.) with documentation. Specifically, here: https://github.com/COBRAPROsimulator/COBRAPRO/blob/157ab80ffda30aa7c51c39cc7a042162efbfbaf0/Examples/Parameter_Identification_Routines/DFN_pso_0_05C.m#L198
- Answer 3: Thank you for this comment. The author will modify the code such that the PSO setting options can be input by the user.

Article Comments

- Comment 4: Presently, the comparison with alternative work in the field is aimed at pure forward modelling packages (PyBaMM, PetLion.jl, LionSimba) which do not aim to solve the inverse problem, a more extensive literature review will find other implementations which are more suitable for comparison.
- Answer 4: Thank you for this comment. The authors focused on comparing DFN open-source codes, since COBRAPRO is a DFN code software. The authors did not include codes using SPM, ESPM, or electric-circuit models in the comparison. The authors were not able to find other DFN open-source codes that include a closed-loop optimization framework other than DEARLIBS. If the reviewer is aware of other DFN codes with a parameter identification framework, please let us know. Thank you. Furthermore, a more detailed study comparing the different DFN codes is included in our accepted JES paper which we will cite in COBRAPRO's GitHub page.

- Comment 5: A table showing the full set of parameters identified for the example would be helpful to summarise the results in one place for the reader.
- Answer 5: Thank you for this comment. The author can indeed include a table showing the full list of DFN parameters. Or, we can cite our JES paper, which has a detailed list of the DFN parameters to save space in the JOSS paper.

- Comment 6: Similarly, a table of correlation metrics for the parameters in the LSA section of the example.
- Answer 6: Thank you for this comment. The authors can add more details on the LSA and correlation analysis results for improved clarity. Also, please note that our JES paper on COBRAPRO provides a detailed account of the LSA and correlation analysis methods.
