### 2024-09-02 #6803 @mbarzegary
@BradyPlanden and @brosaplanella, can you please have a look at the fixes and replies provided by @COBRAPROsimulator and see if they resolve the issues raised by you? If so, I appreciate it if you finalize your reviews.
