### 2024-09-13 712e2c8
print time to solve the DAE to command window
