### 2024-09-13 957bba3
print DAE solver times for each cycle and total cycles to command window
