### 2024-09-14 3a65bb2
include Matlab download in the installation procedure
