+++ README.md
@@ -49,11 +49,13 @@ Installation section below shows how to install the required software.
    ```
    git clone https://github.com/COBRAPROsimulator/COBRAPRO.git
    ```
-2. Download [SUNDIALS 2.6.2](https://computing.llnl.gov/sites/default/files/inline-files/sundials-2.6.2.tar.gz) and unzip the folder. Relocate the sundials-2.6.2 folder inside the COBRAPRO folder.
+2. Download [MATLAB](https://www.mathworks.com/downloads/) if not installed already. Make sure to select both the Global Optimization Toolbox and Parallel Computing Toolbox during the installation process.
+   
+3. Download [SUNDIALS 2.6.2](https://computing.llnl.gov/sites/default/files/inline-files/sundials-2.6.2.tar.gz) and unzip the folder. Relocate the sundials-2.6.2 folder inside the COBRAPRO folder.
 
-3. Download the latest version of [CasADi (MATLAB version)](https://web.casadi.org/get/) corresponding to your operating system. Unzip and move your CasADi folder inside the COBRAPRO folder. Your COBRAPRO folder should now contain the sundials-2.6.2 and CasADi folders.
+4. Download the latest version of [CasADi (MATLAB version)](https://web.casadi.org/get/) corresponding to your operating system. Unzip and move your CasADi folder inside the COBRAPRO folder. Your COBRAPRO folder should now contain the sundials-2.6.2 and CasADi folders.
 
-4. Before we can install SUNDIALS, the following software are required to compile the mex files that will interface with the SUNDIALS IDA solver:
+5. Before we can install SUNDIALS, the following software are required to compile the mex files that will interface with the SUNDIALS IDA solver:
    - __Mac users__: Download [Xcode](https://developer.apple.com/xcode/) application (can be downloaded from Apple’s App Store). Once Xcode[^1] is installed, proceed to accept the license agreement. This can be done by opening the Xcode application, which will launch a license agreement window and click the “Agree” icon, or type
      ```
      sudo xcodebuild -license accept
@@ -61,7 +63,7 @@ Installation section below shows how to install the required software.
      in Terminal. If the license is not accepted, MATLAB may give an error such as “Xcode is installed, but its license has not been accepted”.
    - __Window users__: Download [MinGW](https://www.mathworks.com/matlabcentral/fileexchange/52848-matlab-support-for-mingw-w64-c-c-fortran-compiler)
      
-5. Now you are ready to run `install_COBRAPRO.m`[^2] located inside the COBRAPRO folder. Run `install_COBRAPRO.m` in MATLAB and respond to the prompts displayed in the Command Window in the following manner:
+6. Now you are ready to run `install_COBRAPRO.m`[^2] located inside the COBRAPRO folder. Run `install_COBRAPRO.m` in MATLAB and respond to the prompts displayed in the Command Window in the following manner:
 
 ```
 MEX files will be compiled and built using the above options
@@ -91,7 +93,7 @@ Type the name of your CasADi folder (case-sensitive):
 ```
 &rarr; Type the name of the CasADi folder exactly as it appears and hit enter
 
-6. Successful installation will output to the Command Window:
+7. Successful installation will output to the Command Window:
 ```
 COBRAPRO installed successfully! To get started, try running scripts in the Examples folder.
 ```
