### 2024-09-14 e8d88f7
include missing procedure in sundials installation
