### 2024-09-14 cebb527
Update messages printed to command window
