### 2024-09-14 eed68c3
print unsuccessful initialization if param.sim_print is true
