### 2024-09-14 e9bf541
update objective function messages printed to screen
