+++ paper/paper.md
@@ -28,11 +28,11 @@ COBRAPRO (**Co**-simulation **B**atte**r**y Modeling for **A**ccelerated **P**a*
 
 # Statement of need
 
-Even though parameter calibration is required to accurately predict the dynamical behavior of actual batteries, current DFN modeling tools lack the capability to perform parameter identification [@sulzer_python_2021] [@torchio_lionsimba_2016] [@smith_multiphase_2017] [@berliner_methods_2021]. 
+Even though parameter calibration is required to accurately predict the dynamical behavior of actual batteries, current DFN modeling tools lack the capability to perform parameter identification. 
 
 COMSOL Multiphysics&copy; [@comsol] is a commercially available finite element modeling software commonly used to simulate the DFN model. Although COMSOL lacks a built-in parameter identification feature, it was demonstrated that COMSOL's *LiveLink&trade; for MATLAB&copy;* can be used to establish communication between COMSOL and MATLAB for parameter optimization  [@pozzato_general_2023]. This framework allows users to leverage the versatile suite of optimizers in MATLAB while running COMSOL to generate the model output. However, the expensive licensing fee and proprietary nature of COMSOL create barriers to public access, limiting collaboration and code reproducibility.
 
-In contrast, several open-source DFN model simulation tools have emerged, such as PyBaMM [@sulzer_python_2021], LIONSIMBA [@torchio_lionsimba_2016], PETLION [@berliner_methods_2021], DEARLIBS [@lee_robust_2021], fastDFN[@fastDFN], and MPET [@smith_multiphase_2017]. Among these packages, DEARLIBS is the only software equipped for closed-loop parameter identification using experimental data. Other packages resort to literature-derived parameter values and lack the ability to predict real battery behavior. Taking inspiration from DEARLIBS, COBRAPRO aims to address three primary challenges in the DFN model:
+In contrast, several open-source DFN model simulation tools have emerged, such as PyBaMM [@sulzer_python_2021], LIONSIMBA [@torchio_lionsimba_2016], PETLION [@berliner_methods_2021], DEARLIBS [@lee_robust_2021], fastDFN [@fastDFN], and MPET [@smith_multiphase_2017]. Among these packages, DEARLIBS is the only software equipped with a closed-loop parameter identification routine. Other open-source pacakges operate as open-loop software, requiring known parameters to run models that can accurately predict the behavior of real batteries. However, since open-loop software lack parameter optimization capabilities, they generally rely on uncalibrated, experimentally measured parameters reported in the literature. Taking inspiration from DEARLIBS, COBRAPRO aims to address three primary challenges in the DFN model:
 
 ## Challenge 1. Computational complexity 
 - **Issue:** The DFN model is also known as the pseudo-two-dimensional (P2D) model due to the coupling of the cell thickness (x-direction) and radial particle (r-direction) dimensions. This coupling of dimensions contributes to the model's computational complexity.
