### 2024-09-14 877f3c3
Update paper.bib
