+++ paper/paper.bib
@@ -179,8 +179,9 @@
 @article{ha_cobrapro_2024,
   title = {COBRAPRO: An Open-source Software for the Dolye-Fuller-Newman Model with Co-simulation Parameter Optimization Framework},
   author = {Ha, Sara and Onori, Simona},
-  journal = {Journal of The Electrochemical Society, under review},
+  journal = {Journal of The Electrochemical Society},
   year = {2024},
+  doi = {10.1149/1945-7111/ad7292}
 }
 
 @article{arunachalam_full_2019,
