+++ paper/paper.md
@@ -36,11 +36,11 @@ In contrast, several open-source DFN model simulation tools have emerged, such a
 
 ## Challenge 1. Computational complexity 
 - **Issue:** The DFN model is also known as the pseudo-two-dimensional (P2D) model due to the coupling of the cell thickness (x-direction) and radial particle (r-direction) dimensions. This coupling of dimensions contributes to the model's computational complexity.
-- **Solution:** COBRAPRO leverages the SUNDIALS IDA solver to enable fast computation speed compared to DEARLIBS. At 1C discharge for 10 discretized points in each domain of the cell ($N_p=N_s=N_n=N_{rp},N_{rn}$), COBRAPRO solves the DFN model in 0.708 seconds, while DEARLIBS takes 2.54 minutes, which is a two orders of magnitude improvement (~257 times). Under identifical simulation conditions, LIONSIMBA and PyBaMM computed the model in 1.13 seconds and 0.237 seconds, respectively, demonstrating comparable performance to COBRAPRO. For larger discretization points, COBRAPRO exhibits up to three orders of magnitude improvement in computation speed compared to DEARLIBS. Refer to [@ha_cobrapro_2024] for more details on the computation comparison study between COBRAPRO and other open-source DFN codes.
+- **Solution:** COBRAPRO leverages the SUNDIALS IDA solver to enable fast computation speed compared to DEARLIBS. At 1C discharge for 10 discretized points in each domain of the cell ($N_p=N_s=N_n=N_{rp}=N_{rn}=10$), COBRAPRO solves the DFN model in 0.708 seconds, while DEARLIBS takes 2.54 minutes, which is a two orders of magnitude improvement (~257 times). Under identifical simulation conditions, LIONSIMBA and PyBaMM computed the model in 1.13 seconds and 0.237 seconds, respectively, demonstrating comparable performance to COBRAPRO. For larger discretization points, COBRAPRO exhibits up to three orders of magnitude improvement in computation speed compared to DEARLIBS. Refer to [@ha_cobrapro_2024] for more details on the computation comparison study between COBRAPRO and other open-source DFN codes.
 
 ## Challenge 2. Consistent initial conditions
 - **Issue:** The partial differential equations (PDEs) governing the DFN model are discretized in x- and r-directions to form a system of ordinary differential equations (ODEs) and algebraic equations (AEs), referred to as differential-algebraic equations (DAEs). To solve the DAE system, the correct initial conditions of the AEs are required, which are typically not known *a priori* for the DFN model. Inconsistent initial conditions result in either a failure to start the simulation or the model diverging towards an incorrect solution [@methekar_perturbation_2011].
-- **Solution:** The single step approach [@lawder_extending_2015], a robust initialization method, is implemented in COBRAPRO that determines the initial conditions by modifying the AEs into implicit ODEs and seamlessly integrates the consistent initial conditions to solve the DFN model. As shown in [@ha_cobrapro_2024], the single-step approach is able to determine consistent initial conditions more robustly than using a typical nonlinear algebraic solver, such as SUNDIALS IDA CalcIC or CasADI's rootfinder function, which are used in the other open-source codes. 
[…]
