### 2024-09-16 cc8a30b
include more user setting options for pso
