### 2024-09-16 57495d6
include more pso setting options for users
