+++ Examples/Parameter_Identification_Routines/DFN_pso_HPPC.m
@@ -79,10 +79,39 @@ lower_bounds.Dsn = 10^(log10(param.Dsn)*(1+pct));
 upper_bounds.Dsn = 10^(log10(param.Dsn)*(1-pct));
 
 %--------------------------------------------------------------------------
-% Enter number of particles for PSO
+% PSO Settings: Enter number of particles for PSO
 % (for code testing purposes, reduce particle number,e.g., particle_num = 2)
 %--------------------------------------------------------------------------
 particle_num = 100;
+
+%--------------------------------------------------------------------------
+% PSO Settings: Enter PSO exit condition 
+% (https://www.mathworks.com/help/gads/particleswarm.html#budidgf-options)
+%--------------------------------------------------------------------------
+% ----------------------------------
+% Option 1: Maximum iterations
+% ----------------------------------
+% exit_type = 'MaxIterations' -> PSO exits after maximum number of PSO iterations reached
+% max_iterations = ... -> Maximum number of PSO iterations
+% ----------------------------------
+% Option 2: Maximum stall iterations
+% ----------------------------------
+% exit_type = 'MaxStallIterations' -> PSO iterations end when the relative change in best objective function value over the last MaxStallIterations iterations is less than FunctionTolerance
+% max_iterations = ... -> Number of MaxStallIterations
+% max_stall_tolerance = ... -> FunctionTolerance value
+exit_type = 'MaxStallIterations';
+max_iterations = 5;
+max_stall_tolerance = 0.5e-6;
+
+%--------------------------------------------------------------------------
+% PSO Settings: Enter PSO weight 
+%--------------------------------------------------------------------------
+% References: 
+% [1] https://www.mathworks.com/help/gads/particle-swarm-optimization-algorithm.html
+% [2] G. Pozzato and S. Onori, "A General Matlab and COMSOL Co-simulation Framework for Model Parameter Optimization: Lithium-Ion Battery and Gasoline Particulate Filter Case Studies," presented at the Automotive Technical Papers, Warrendale, Pennsylvania, United States, Jul. 2023, pp. 2023-01–5047. doi: 10.4271/2023-01-5047.
+%--------------------------------------------------------------------------
+social_adjustment = 3.6; 
+self_adjustment = 0.3;  
 
 %--------------------------------------------------------------------------
 % Load Experimental Data 
@@ -213,11 +242,11 @@ options = optimoptions('particleswarm',...
                        'UseVectorized',true,...
                        'SwarmSize',particle_num,...
                        'Display','iter',...
-                       'MaxStallIterations',5,...
-                       'FunctionTolerance',0.5e-6,...
+                       exit_type,max_iterations,...
+                       'FunctionTolerance',max_stall_tolerance,...
                        'InitialSwarmMatrix',init_pos,...
-                       'SocialAdjustmentWeight',3.6,... 
-                       'SelfAdjustmentWeight',0.3);  
+                       'SocialAdjustmentWeight',social_adjustment,... 
+                       'SelfAdjustmentWeight',self_adjustment); 
 
 %% RUN PSO 
 fprintf('\nStart running PSO!...\n')
