### 2024-09-16 1dadaee
include JES paper citation
