+++ DFN_parameters/INR21700_M50T/Un_Graphite_Si_Chen2020.m
@@ -1,4 +1,4 @@
-function Un = Un_Graphite_Si(theta)
+function Un = Un_Graphite_Si_Chen2020(theta)
 % Un_Graphite_Si is the experimentally determined open-circuit potential (OCP)
 % of the Graphite-Silicon anode from [1]. OCP was determined from half-cell GITT
 % tests in [1].
