### 2024-09-16 7ec56ba
Update and rename DFN_parameters/Parameters_LG_INR21700_M50.m to DFN_parameters/LG_INR21700_M50/Parameters_LG_INR21700_M50.m
