### 2024-09-16 ed6c525
Update and rename DFN_parameters/Up_NMC811.m to DFN_parameters/LG_INR21700_M50/Up_NMC811_Chen2020.m
