### 2024-09-16 e09c6c7
Update and rename DFN_parameters/LG_INR21700_M50/Kappa_eff_Chen2020.m to parameters/Kappa_eff_Nyman2008.m
