### 2024-09-16 faadef9
Update and rename DFN_parameters/lambda_Kremer.m to parameters/electrolyte_functional/lambda_Kremer2020.m
