+++ parameters/electrolyte_functional/t1_Kremer2020.m
@@ -1,4 +1,4 @@
-function t1 = t1_Kremer(ce, param)
+function t1 = t1_Kremer2020(ce, param)
 % t1_Kremer evaluates the transference number as a function of electrolyte
 % concentration (ce). Function from [1].
 %
