### 2024-09-16 8f2034f
Update and rename DFN_parameters/t1_Nyman.m to parameters/electrolyte_functional/t1_Nyman2008.m
