### 2024-09-16 c5259a7
Update Parameters_LG_INR21700_M50.m
