### 2024-09-16 8fa36a7
Update paper.bib
