+++ paper/paper.bib
@@ -269,3 +269,26 @@
   url = {https://github.com/scott-moura/fastDFN},
 }
 
+@article{gardner2022sundials,
+	title     = {Enabling new flexibility in the {SUNDIALS} suite of nonlinear and differential/algebraic equation solvers},
+	author    = {Gardner, David J and Reynolds, Daniel R and Woodward, Carol S and Balos, Cody J},
+	journal   = {ACM Transactions on Mathematical Software (TOMS)},
+	publisher = {ACM},
+	volume    = {48},
+	number    = {3},
+	pages     = {1--24},
+	year      = {2022},
+	doi       = {10.1145/3539801}
+}
+
+@article{hindmarsh2005sundials,
+	title     = {{SUNDIALS}: Suite of nonlinear and differential/algebraic equation solvers},
+	author    = {Hindmarsh, Alan C and Brown, Peter N and Grant, Keith E and Lee, Steven L and Serban, Radu and Shumaker, Dan E and Woodward, Carol S},
+	journal   = {ACM Transactions on Mathematical Software (TOMS)},
+	publisher = {ACM},
+	volume    = {31},
+	number    = {3},
+	pages     = {363--396},
+	year      = {2005},
+	doi       = {10.1145/1089014.1089020}
+}
