+++ paper/paper.md
@@ -40,7 +40,7 @@ In contrast, several open-source DFN model simulation tools have emerged, such a
 
 ## Challenge 2. Consistent initial conditions
 - **Issue:** The partial differential equations (PDEs) governing the DFN model are discretized in x- and r-directions to form a system of ordinary differential equations (ODEs) and algebraic equations (AEs), referred to as differential-algebraic equations (DAEs). To solve the DAE system, the correct initial conditions of the AEs are required, which are typically not known *a priori* for the DFN model. Inconsistent initial conditions result in either a failure to start the simulation or the model diverging towards an incorrect solution [@methekar_perturbation_2011].
-- **Solution:** The single-step approach [@lawder_extending_2015], a robust initialization method implemented in COBRAPRO, modifies the AEs into implicit ODEs to determine consistent initial conditions. This approach eliminates the need for nonlinear algebraic solvers, such as SUNDIALS IDACalcIC or CasADi's rootfinder, which are commonly used by other open-source DFN codes but have been shown to have convergence issues [@methekar_perturbation_2011, @lawder_extending_2015]. In [@ha_cobrapro_2024], a comparison study between the single-step approach and SUNDIALS IDA CalcIC demonstrates that the single-step approach consistently determines initial conditions across different C-rates and node numbers. For details on the implementation of the single-step approach, see [@ha_cobrapro_2024].
+- **Solution:** The single-step approach [@lawder_extending_2015], a robust initialization method implemented in COBRAPRO, modifies the AEs into implicit ODEs to determine consistent initial conditions. This approach eliminates the need for nonlinear algebraic solvers, such as SUNDIALS IDACalcIC or CasADi's rootfinder, which are commonly used by other open-source DFN codes but have been shown to have convergence issues [@methekar_perturbation_2011][@lawder_extending_2015]. In [@ha_cobrapro_2024], a comparison study between the single-step approach and SUNDIALS IDA CalcIC demonstrates that the single-step approach consistently determines initial conditions across different C-rates and node numbers. For details on the implementation of the single-step approach, see [@ha_cobrapro_2024].
 
 ## Challenge 3. Unknown model parameters
 - **Issue:** As highlighted earlier, battery parameters are frequently unknown, and even if obtained through experimental characterization, parameter calibration is essential to accurately simulate the DFN model.
