+++ paper/paper.bib
@@ -278,7 +278,7 @@
 	number    = {3},
 	pages     = {1--24},
 	year      = {2022},
-	doi       = {10.1145/3539801}
+	doi       = {10.1145/3539801},
 }
 
 @article{hindmarsh2005sundials,
@@ -290,5 +290,5 @@
 	number    = {3},
 	pages     = {363--396},
 	year      = {2005},
-	doi       = {10.1145/1089014.1089020}
+	doi 	  = {10.1145/1089014.1089020},
 }
