### 2024-09-17 e3a75a1
Update paper.bib
