+++ paper/paper.bib
@@ -292,3 +292,15 @@
 	year      = {2005},
 	doi 	  = {10.1145/1089014.1089020},
 }
+
+@article{li_datadriven_2022,
+  title = {Data-Driven Systematic Parameter Identification of an Electrochemical Model for Lithium-Ion Batteries with Artificial Intelligence},
+  author = {Li, Weihan and Demir, Iskender and Cao, Decheng and J{\"o}st, Dominik and Ringbeck, Florian and Junker, Mark and Sauer, Dirk Uwe},
+  year = {2022},
+  month = jan,
+  journal = {Energy Storage Materials},
+  volume = {44},
+  pages = {557--570},
+  issn = {24058297},
+  doi = {10.1016/j.ensm.2021.10.023},
+}
