+++ paper/paper.bib
@@ -304,3 +304,29 @@
   issn = {24058297},
   doi = {10.1016/j.ensm.2021.10.023},
 }
+
+@article{xu_comparative_2023,
+  title = {Comparative {{Analysis}} of {{Numerical Methods}} for {{Lithium-Ion Battery Electrochemical Modeling}}},
+  author = {Xu, Le and Cooper, Julian and Allam, Anirudh and Onori, Simona},
+  year = {2023},
+  month = dec,
+  journal = {Journal of The Electrochemical Society},
+  volume = {170},
+  number = {12},
+  pages = {120525},
+  issn = {0013-4651, 1945-7111},
+  doi = {10.1149/1945-7111/ad1293},
+}
+
+@article{allam_online_2021,
+  title = {Online {{Capacity Estimation}} for {{Lithium-Ion Battery Cells}} via an {{Electrochemical Model-Based Adaptive Interconnected Observer}}},
+  author = {Allam, Anirudh and Onori, Simona},
+  year = {2021},
+  month = jul,
+  journal = {IEEE Transactions on Control Systems Technology},
+  volume = {29},
+  number = {4},
+  pages = {1636--1651},
+  issn = {1063-6536, 1558-0865, 2374-0159},
+  doi = {10.1109/TCST.2020.3017566},
+}
