+++ paper/paper.bib
@@ -330,3 +330,29 @@
   issn = {1063-6536, 1558-0865, 2374-0159},
   doi = {10.1109/TCST.2020.3017566},
 }
+
+@article{goshtasbi_effective_2020,
+	title = {Effective {{Parameterization}} of {{PEM Fuel Cell Models}}---{{Part I}}: {{Sensitivity Analysis}} and {{Parameter Identifiability}}},
+	shorttitle = {Effective {{Parameterization}} of {{PEM Fuel Cell Models}}---{{Part I}}},
+	author = {Goshtasbi, Alireza and Chen, Jixin and Waldecker, James R. and Hirano, Shinichi and Ersal, Tulga},
+	year = {2020},
+	month = mar,
+	journal = {Journal of The Electrochemical Society},
+	volume = {167},
+	number = {4},
+	pages = {044504},
+	issn = {0013-4651, 1945-7111},
+	doi = {10.1149/1945-7111/ab7091},
+}
+
+@article{forman_genetic_2012,
+  title = {Genetic Identification and Fisher Identifiability Analysis of the {{Doyle}}{\textendash}{{Fuller}}{\textendash}{{Newman}} Model from Experimental Cycling of a {{LiFePO4}} Cell},
+  author = {Forman, Joel C. and Moura, Scott J. and Stein, Jeffrey L. and Fathy, Hosam K.},
+  year = {2012},
+  month = jul,
+  journal = {Journal of Power Sources},
+  volume = {210},
+  pages = {263--275},
+  issn = {03787753},
+  doi = {10.1016/j.jpowsour.2012.03.009},
+}
