+++ paper/paper.md
@@ -39,12 +39,12 @@ In contrast, several open-source DFN model simulation tools have emerged, such a
 - **Solution:** COBRAPRO leverages the SUNDIALS IDA solver [@hindmarsh2005sundials], [@gardner2022sundials] to enable fast computation speed compared to DEARLIBS, which uses MATLAB&copy;'s Symbolic Math Toolbox and ode15 solver. For 1C discharge given 10 discretized points in the electrodes, separator, and particle domains ($N_p=N_s=N_n=N_{rp}=N_{rn}=10$), COBRAPRO solves the DFN model in ~0.708 seconds, while DEARLIBS takes ~2.54 minutes, which is a two orders of magnitude improvement (~257 times). Under identifical simulation conditions, LIONSIMBA and PyBaMM computed the model in ~1.13 seconds and ~0.237 seconds, respectively, demonstrating comparable performance to COBRAPRO. For larger discretization points, COBRAPRO exhibits up to three orders of magnitude improvement in computation speed compared to DEARLIBS. Refer to [@ha_cobrapro_2024] for more details on the computation comparison study between COBRAPRO and other open-source DFN codes.
 
 ## Challenge 2. Determining consistent initial conditions
-- **Issue:** The partial differential equations (PDEs) governing the DFN model are discretized in x- and r-directions to form a system of ordinary differential equations (ODEs) and algebraic equations (AEs), referred to as differential-algebraic equations (DAEs). To solve the DAE system, the correct initial conditions of the AEs are required, which are typically not known *a priori* for the DFN model [@methekar_perturbation_2011],[@lee_robust_2021]. Inconsistent initial conditions result in either a failure to start the simulation or the model diverging towards an incorrect solution [@methekar_perturbation_2011].
-- **Solution:** The single-step approach [@lawder_extending_2015], a robust initialization method implemented in COBRAPRO and DEARLIBS, modifies the AEs into implicit ODEs to determine consistent initial conditions. This approach eliminates the need for nonlinear algebraic solvers, such as SUNDIALS IDACalcIC or CasADi's rootfinder, which are commonly used by other open-source DFN codes and shown to exhibit convergence issues [@methekar_perturbation_2011]. In [@ha_cobrapro_2024], a comparison study between the single-step approach and SUNDIALS IDACalcIC demonstrates that the single-step approach consistently determines initial conditions across different C-rates and node numbers. For more details, refer to [@ha_cobrapro_2024].
+- **Issue:** The partial differential equations (PDEs) governing the DFN model are discretized in x- and r-directions to form a system of ordinary differential equations (ODEs) and algebraic equations (AEs), referred to as differential-algebraic equations (DAEs). To solve the DAE system, it is essential to provide the correct initial conditions for the algebraic variables, which are typically not known *a priori* and must satisfy the AEs in the DFN model [@lee_robust_2021]. Inconsistent initial conditions result in either a failure to start the simulation or the model diverging towards an incorrect solution [@methekar_perturbation_2011].
+- **Solution:** The single-step approach [@lawder_extending_2015], a robust initialization method implemented in COBRAPRO and DEARLIBS, modifies the AEs into implicit ODEs to determine consistent initial conditions. This approach eliminates the need for nonlinear algebraic solvers, such as SUNDIALS IDACalcIC or CasADi's rootfinder, which are commonly used in other open-source DFN codes but have been shown to experience convergence issues [@methekar_perturbation_2011]. In [@ha_cobrapro_2024], a comparison study between the single-step approach and SUNDIALS IDACalcIC demonstrates that the single-step approach consistently determines initial conditions across different C-rates and node numbers, outperforming SUNDIALS IDACalcIC. For more details, refer to [@ha_cobrapro_2024].
 
[…]
