+++ Examples/Local_Sensitivity_Analysis/DFN_LSA_Corr_HPPC.m
@@ -37,10 +37,10 @@ param.upperCutoffVoltage = 4.3;
 % Enter names of parameters to conduct LSA (make sure names match the
 % parameter names in "param" structure containing nominal parameters)
 %--------------------------------------------------------------------------
-param_LSA_HPPC = {'Dsp' 'Dsn' 't1_constant' 'kp' 'kn' 'c0' 'De' 'Kappa'};
+param_LSA_HPPC = {'Rc' 'Dsp' 'Dsn' 'kp' 'kn' 'De' 'Kappa' 'sigmap' 'sigman'};
 
 % Type in latex version of param_LSA_HPPC for nice plots
-theta_names = {'$D_{s,p}$' '$D_{s,n}$' '$t_{+}$' '$k_p$' '$k_n$' '$c_0$' '$D_e$' '$\kappa$'};
+theta_names = {'$R_c$' '$D_{s,p}$' '$D_{s,n}$' '$k_p$' '$k_n$' '$D$' '$\kappa$' '$\sigma_p$' '$\sigma_n$'};
 
 %--------------------------------------------------------------------------
 % Perturbation coefficient for LSA [-]
@@ -48,20 +48,13 @@ theta_names = {'$D_{s,p}$' '$D_{s,n}$' '$t_{+}$' '$k_p$' '$k_n$' '$c_0$' '$D_e$'
 pct=0.05;
 
 %--------------------------------------------------------------------------
-% Define sensitivity index threshold
-%--------------------------------------------------------------------------
-% Parameters with sensitivity index higher than beta_LSA will be considered
-% the "identifiable" parameters for the HPPC profile
-%--------------------------------------------------------------------------
-beta_LSA = 0.5;
-
-%--------------------------------------------------------------------------
 % Define correlation coefficient threshold
 %--------------------------------------------------------------------------
 % Correlation coefficient threshold used to determine the uncorrelated parameter  
-% vector, which will be considered the "identifiable" parameters for the HPPC profile
+% vector prioritized by sensitivity, which will be considered the "identifiable" parameters for the HPPC profile
+% Input vector to yield multiple identifiable parameter sets for each correlation threshold
 %--------------------------------------------------------------------------
-beta_corr = 0.99;
+beta_corr = [0.8 0.9 0.95 0.98 0.99];
 
 %--------------------------------------------------------------------------
 % Load HPPC profile input current vector [A] and simulation time [s] 
@@ -218,7 +211,6 @@ plot(S_2norm_V_SOC_all_sort,'o','MarkerSize',12,'linewidth',2,'DisplayName','$S^
 plot(S_2norm_V_all_sort,'^','MarkerSize',12,'linewidth',2,'DisplayName','$S^{V}$')
 xticks(1:length(theta_names));xticklabels(theta_names(ind));
 ylabel('Sensitivity [-]')
-set(gca, 'YScale', 'log'); 
 title('HPPC Sensitivity');%title('2-norm of Normalized Sensitivity Vector')
 set(gcf,'color','w');legend('location','best')
 set(findall(gcf,'-property','interpreter'),'interpreter','latex')
@@ -227,12 +219,22 @@ set(gcf, 'Units','inches','Position', [1 1 14 8]);set(gca,'Fontsize',32);
 
 %% Correlation matrix calculation
 
- min_length_V_SOC_vec = min(length_V_SOC_vec);
+% Find indices with non-zero sensitivity 
[…]
