+++ Examples/Parameter_Identifiability_Analysis/DFN_LSA_Corr_CC.m
@@ -24,9 +24,14 @@
 clc;clear;close all
 
 %--------------------------------------------------------------------------
-% Load nominal/reference parameters
+% Option 1: Enter identified parameters
 %--------------------------------------------------------------------------
-param = Parameters_LG_INR21700_M50;
+load('identified_parameters_0_05C.mat','param')
+param.upperCutoffVoltage = 4.3; 
+%--------------------------------------------------------------------------
+% Option 2: If you don't have identified parameters, then load parameters from literature [Chen 2020]
+%--------------------------------------------------------------------------
+% param = Parameters_LG_INR21700_M50;
 
 %--------------------------------------------------------------------------
 % Enter names of parameters to conduct LSA (make sure names match the
@@ -43,20 +48,12 @@ theta_names = {'$\varepsilon_p$' '$\varepsilon_n$' '$R_p$' '$R_n$'};
 pct=0.05;
 
 %--------------------------------------------------------------------------
-% Define sensitivity index threshold
-%--------------------------------------------------------------------------
-% Parameters with sensitivity index higher than beta_LSA will be considered
-% the "identifiable" parameters for the HPPC profile
-%--------------------------------------------------------------------------
-beta_LSA = 0.5;
-
-%--------------------------------------------------------------------------
 % Define correlation coefficient threshold
 %--------------------------------------------------------------------------
-% Correlation coefficient threshold used to determine the uncorrelated parameter  
-% vector, which will be considered the "identifiable" parameters for the HPPC profile
+% If beta_corr is a vector, then the script will calculate the identifiable
+% parameter set for each correlation coefficient threshold in beta_corr
 %--------------------------------------------------------------------------
-beta_corr = 0.8;
+beta_corr = [0.8 0.9 0.95 0.98 0.99];
 
 %--------------------------------------------------------------------------
 % Define input current density [A.m-2]
@@ -217,12 +214,22 @@ set(gcf, 'Units','inches','Position', [1 1 15 8]);set(gca,'Fontsize',32);
 
 %% Correlation matrix calculation
 
- min_length_V_SOC_vec = min(length_V_SOC_vec);
+% Find indices with non-zero sensitivity 
+nonzero_S_2norm_V_SOC_ind = S_2norm_V_SOC_all~=0;
+% Remove parameters with zero sensitivity for correlation matrix calculation
+length_V_SOC_vec_new = length_V_SOC_vec(nonzero_S_2norm_V_SOC_ind);
+param_LSA_HPPC_new = param_LSA_HPPC(nonzero_S_2norm_V_SOC_ind);
+theta_names_new = theta_names(nonzero_S_2norm_V_SOC_ind);
+S_norm_V_SOC_new = S_norm_V_SOC(nonzero_S_2norm_V_SOC_ind);
+S_2norm_V_SOC_all_new = S_2norm_V_SOC_all(nonzero_S_2norm_V_SOC_ind);
 
- S_2norm_V_SOC_matrix = zeros(min_length_V_SOC_vec,length(theta_names));
- for i=1:length(theta_names)
-     S_2norm_V_SOC_matrix(:,i) = S_norm_V_SOC{i}(1:min_length_V_SOC_vec,1);
+min_length_V_SOC_vec = min(length_V_SOC_vec_new);
+
+ S_2norm_V_SOC_matrix = zeros(min_length_V_SOC_vec,length(theta_names_new));
+ for i=1:length(theta_names_new)
[…]
