+++ Examples/Parameter_Identification_Results/DFN_pso_HPPC_identification.m
@@ -27,7 +27,7 @@ clc;clear;close all
 %--------------------------------------------------------------------------
 % Enter your identified parameters 
 %--------------------------------------------------------------------------
-load('identified_parameters_HPPC_0_99Corr.mat')
+load('identified_parameters_HPPC_0_95Corr.mat')
 
 %--------------------------------------------------------------------------
 % Print simulation time 
