### 2024-09-19 b3a8d22
Update paper.md
