+++ paper/paper.md
@@ -77,7 +77,7 @@ View COBRAPRO's [README on Github](https://github.com/COBRAPROsimulator/COBRAPRO
 
 ## C/20 Capacity Test Identification
 
-In `DFN_pso_0_05C.m`, the `User Input` section is used to define the parameter names, PSO settings, experimental data, etc. A preview of the `User Input` section is provided here.
+In `Examples/Parameter_Identification_Routines/DFN_pso_0_05C.m`, the `User Input` section is used to define the parameter names, PSO settings, experimental data, etc. A preview of the `User Input` section is provided here.
 
 First, load the `Parameters_LG_INR21700_M50.m` function, which outputs a `param` structure containing the nominal (experimentally measured) parameters of a LG INR21700-M50 cell [@chen_development_2020], as well as the DFN simulation settings, e.g., discretization method, method to determine consistent initial condition, constant or variable current type, etc.:
 ```MATLAB
@@ -158,14 +158,14 @@ J_tot =0.003833 [-]
 ```
 The code also plots the simulation results generated from the identified parameters and the experimental data, as shown in \autoref{fig:V_0_05C} and \autoref{fig:SOC_0_05C}. 
 
-To replicate the results, run the code `Examples/Parameter_Identification_Results/DFN_pso_0_05C_identification.m`, which loads the identified parameters determined from this case study. 
+Run `Examples/Parameter_Identification_Routines/DFN_pso_0_05C_identification.m` to replicate the results shown here (code will load the identified parameters shown in this case study). 
 
 ![C/20 capacity test voltage identification results.\label{fig:V_0_05C}](0_05C_Voltage.png){ width=65% }
 
 ![C/20 capacity test positive and negative electrode SOC identification results.\label{fig:SOC_0_05C}](0_05C_SOC.png){ width=65% }
 
 ## Parameter Identifiability Analysis
-In this step, parameter identifiability analysis is conducted to determine a subset of identifiable parameters given the HPPC profile in `DFN_LSA_Corr_HPPC.m`. In the `User Input` section, load the `param` structure that contains the identified stoichiometric parameter values from `DFN_pso_0_05C.m`:
+In `Examples/Parameter_Identifiability_Analysis/DFN_LSA_Corr_HPPC.m`, parameter identifiability analysis is conducted to determine a subset of identifiable parameters given the HPPC profile. In the `User Input` section, load the `param` structure that contains the identified stoichiometric parameter values from `DFN_pso_0_05C.m`:
 ```MATLAB
 %% User Input
 % Load nominal parameters and identified stoichiometric parameters
@@ -200,8 +200,9 @@ pct=0.05;
 % Define correlation coefficient threshold
 %--------------------------------------------------------------------------
 % Correlation coefficient threshold used to determine the uncorrelated parameter  
-% vector prioritized by sensitivity, which will be considered the "identifiable" parameters for the HPPC profile
-% Input vector to yield multiple identifiable parameter sets for each correlation threshold
+% vector prioritized by sensitivity, which will be considered the "identifiable"
+% parameters for the HPPC profile (input vector to yield multiple identifiable
+% parameter sets for each correlation threshold)
 %--------------------------------------------------------------------------
 beta_corr = [0.8 0.9 0.95 0.98 0.99];
 ```
@@ -227,8 +228,8 @@ Rc Dsp Dsn kp De (threshold = 0.99)
 ```
 
 ## HPPC Identification
-The `DFN_pso_HPPC.m` file's `User Input` section is similar to the one described in `DFN_pso_0_05C.m`.
[…]
