### 2024-09-19 #6803 @COBRAPROsimulator
Hello, @mbarzegary, @brosaplanella, @BradyPlanden, thank you again for your valuable comments. We wanted to let you know what we are in the process of making those changes, in the code and the paper. When complete, we will respond with a detailed list of the changes we made to address your comments. Thank you.
