### 2024-09-22 0bf9269
Update test_2_comsolValidation.m
