### 2024-09-22 COBRAPROsimulator/COBRAPRO#1 @COBRAPROsimulator
Hi @yuefan98, upon further investigation, I am re-visiting this thread.

For Macs with Apple silicon processor, I noticed the issue is with the CasADi version. There are two versions of the Mac M1 version of [CasADi for MATLAB](https://web.casadi.org/get/): Option 1. R2020b or later and Option 2. R2018b or later. When I use the first option (R2020b or later), I get the same error you got "Unrecognized variable or function  <img width="920" alt="Screenshot 2024-09-22 at 1 48 05 PM" src="https://github.com/user-attachments/assets/a6e5be15-8626-4202-8a93-128be501c501">  However, if I use the second version, R2018b or later, CasADi seems to work fine for MATLAB 2023b, 2024a, 2024b (ARM/Apple Silicon versions of MATLAB).

If possible, could you please confirm if this is also true for your Mac with the Apple Silicon chip? Thank you so much.

 you please
