+++ documentation.md
@@ -1,1 +1,5 @@
+layout: page
+title: "Documentation"
+permalink: /documentation
 
+Test
