+++ _config.yml
@@ -1,0 +1,3 @@
+theme: jekyll-theme-architect
+title: COBRAPRO
+description: Bookmark this to keep an eye on my project updates!
