+++ _config.yml
@@ -1,3 +1,3 @@
-theme: jekyll-theme-architect
+theme: jekyll-theme-minimal
 title: COBRAPRO
 description: Bookmark this to keep an eye on my project updates!
