+++ _config.yml
@@ -1,3 +1,3 @@
-theme: jekyll-theme-minimal
+theme: minima
 title: COBRAPRO
 description: Bookmark this to keep an eye on my project updates!
