+++ docs/index.md
@@ -1,0 +1,3 @@
+# Home
+
+This is a test.
