+++ mkdocs.yml
@@ -1,0 +1,3 @@
+site_name: COBRAPRO
+site_url: https://wp-net.github.io/COBRAPRO/
+theme: readthedocs
