+++ mkdocs.yml
@@ -1,3 +1,3 @@
 site_name: COBRAPRO
-site_url: https://wp-net.github.io/COBRAPRO/
+site_url:  https://cobraprosimulator.github.io/COBRAPRO/
 theme: readthedocs
