+++ mkdocs.yml
@@ -1,3 +1,5 @@
 site_name: COBRAPRO
 site_url:  https://cobraprosimulator.github.io/COBRAPRO/
+nav:
+    - Home: index.md
 theme: readthedocs
