+++ .github/workflows/documentation.yml
@@ -25,6 +25,8 @@ jobs:
         with:
           fetch-depth: 0
       - uses: actions/setup-python@v4
+        with:
+          python-version: '3.12'
       - run: pip install --upgrade pip && pip install mkdocs mkdocs-gen-files
       - run: git config user.name 'github-actions[bot]' && git config user.email 'github-actions[bot]@users.noreply.github.com'
       - name: Publish docs
