+++ .github/workflows/documentation.yml
@@ -3,10 +3,7 @@
 name: Docs
 
 # Controls when the workflow will run
-on:
-  # Triggers the workflow on push or pull request events but only for the master branch
-  push:
-    branches: [ master ]
+on: [push]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
