+++ documentation.md
@@ -1,5 +1,0 @@
-layout: page
-title: "Documentation"
-permalink: /documentation
-
-Test
