+++ .github/workflows/documentation.yml
@@ -1,30 +1,28 @@
 # This is a basic workflow to help you get started with Actions
-
 name: Docs
-
 # Controls when the workflow will run
-on: [push]
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
+on:
+# Triggers the workflow on push or pull request events but only for the master branch
+push:
+branches: [ master ]
+# Allows you to run this workflow manually from the Actions tab
+workflow_dispatch:
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
-  # This workflow contains a single job called "build"
-  build:
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
+# This workflow contains a single job called "build"
+build:
+# The type of runner that the job will run on
+runs-on: ubuntu-latest
+# Steps represent a sequence of tasks that will be executed as part of the job
+steps:
+# Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@v2
+with:
+fetch-depth: 0
       - uses: actions/setup-python@v4
         with:
           python-version: '3.12'
       - run: pip install --upgrade pip && pip install mkdocs mkdocs-gen-files
       - run: git config user.name 'github-actions[bot]' && git config user.email 'github-actions[bot]@users.noreply.github.com'
       - name: Publish docs
-        run: mkdocs gh-deploy
+run: mkdocs gh-deploy --force
