+++ .github/workflows/documentation.yml
@@ -16,7 +16,7 @@ runs-on: ubuntu-latest
 # Steps represent a sequence of tasks that will be executed as part of the job
 steps:
 # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
 with:
 fetch-depth: 0
       - uses: actions/setup-python@v4
