### 2024-09-27 COBRAPROsimulator/COBRAPRO#1 @yuefan98
Thanks for checking back on this. It works on my side now with option 2 installed for the Apple Silicon processor. For your reference, the version I am using is the following _casadi-3.6.6-osx_arm64-matlab2018b_
