### 2024-09-30 COBRAPROsimulator/COBRAPRO#1 @COBRAPROsimulator
@yuefan98, thanks so much for confirming. I will update the README file accordingly, notifying that for Mac Silicon users Option 2 should be used for CasADi.
