### 2024-09-30 4d08f6e
include casadi note
