+++ README.md
@@ -66,7 +66,7 @@ Installation section below shows how to install the required software.
 3. Download [SUNDIALS 2.6.2](https://computing.llnl.gov/sites/default/files/inline-files/sundials-2.6.2.tar.gz) and unzip the folder. Relocate the sundials-2.6.2 folder inside the COBRAPRO folder.
 
 4. Download the latest version of [CasADi (MATLAB version)](https://web.casadi.org/get/) corresponding to your operating system. Unzip and move your CasADi folder inside the COBRAPRO folder. Your COBRAPRO folder should now contain the sundials-2.6.2 and CasADi folders.
-- __IMPORTANT NOTE__: For Mac Apple Silicon users, please read the following if you are using the Apple Silicon version of MATLAB. In [CasADi (MATLAB version)](https://web.casadi.org/get/), you will need to download the Mac M1 version. For Mac M1 versions, please download the "R2018b or later", _NOT_ the the "R2020b or later" version. We have noticed that "R2018b or later" is not compatible with Apple Silicon version of MATLAB. This comment is valid as of CasADi v3.6.6.
+   - __IMPORTANT NOTE__: For Mac Apple Silicon users using the Apple Silicon version of MATLAB, note that you will need to download the Mac M1 version for [CasADi (MATLAB version)](https://web.casadi.org/get/). However, lease download the "R2018b or later", _NOT_ the the "R2020b or later" for the Mac M1 version. We have noticed that "R2018b or later" for the Mac M1 version is not compatible with Apple Silicon version of MATLAB. This comment is valid as of CasADi v3.6.6.
 
 6. Before we can install SUNDIALS, the following software are required to compile the mex files that will interface with the SUNDIALS IDA solver:
    - __Mac users__: Download [Xcode](https://developer.apple.com/xcode/) application (can be downloaded from Apple’s App Store). Once Xcode[^1] is installed, proceed to accept the license agreement. This can be done by opening the Xcode application, which will launch a license agreement window and click the “Agree” icon, or type
