+++ docs/index.md
@@ -1,3 +1,0 @@
-# Home
-
-This is a test.
