+++ docs/Guidelines/justlogic.rst
@@ -1,0 +1,17 @@
+Just Logic
+++++++++++
+
+Key Goals - test
+=================
+
+Options
+=======
+
+CodeMaster
+----------
+
+Robot Turtles
+--------------
+
+Primo / Cubetto
+---------------
