+++ docs/Options/justlogic.rst
@@ -1,0 +1,30 @@
+Just Logic
+++++++++++
+
+Key Goals
+=================
+
+.. list-table:: Comparison
+    :widths: 20 10 10 15
+    :header-rows: 1
+    
+    * - Platform
+      - self
+      - Cost
+      - flex
+    * - rasberry
+      - no
+      - $30
+      - limitless
+
+Options
+=======
+
+CodeMaster
+----------
+
+Robot Turtles
+--------------
+
+Primo / Cubetto
+---------------
