+++ docs/index.rst
@@ -1,0 +1,36 @@
+.. COBRAPRO documentation master file, created by
+   sphinx-quickstart on Fri Oct  4 16:27:47 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+COBRAPRO documentation
+======================
+This is the documentation for CORBAPRO. this is a link :doc:`/Options/justcode`
+this is a link to a particular section :ref:`Options section for justcode <codeOptions>`
+
+.. caution:: 
+   note aution
+
+.. tip:: 
+   these are tips
+
+.. note:: 
+   these are notes
+
+.. image:: /images/flowchart_v2.jpg
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+   :caption: Options:
+
+   Options/justlogic
+   Options/justcode
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+   :caption: Guidelines:
+
+   Guidelines/justlogic
+   Guidelines/justcode
