+++ docs/Guidelines/justcode.rst
@@ -1,0 +1,6 @@
+justcode
+========
+
+Key Goals
+---------
+
