+++ docs/Options/justcode.rst
@@ -1,0 +1,12 @@
+justcode
+========
+
+Key Goals
+---------
+.. _codeOptions:
+
+Here is some code::
+
+    function y = makey(x)
+        y = x+1
+    end
