+++ .readthedocs.yaml
@@ -8,7 +8,10 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.8"
+    python: "3.12"
+    nodejs: "18"
+    rust: "1.64"
+    golang: "1.19"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
@@ -23,7 +26,7 @@ sphinx:
   # You can configure Sphinx to use a different builder, for instance use the dirhtml builder for simpler URLs
   # builder: "dirhtml"
   # Fail on all warnings to avoid broken references
-  # fail_on_warning: true
+  fail_on_warning: true
 
 # Optionally build your docs in additional formats such as PDF and ePub
 # formats:
