+++ .readthedocs.yaml
@@ -8,7 +8,7 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.12"
+    python: "3.8"
     nodejs: "18"
     rust: "1.64"
     golang: "1.19"
