+++ .readthedocs.yaml
@@ -8,10 +8,7 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3.8"
-    nodejs: "18"
-    rust: "1.64"
-    golang: "1.19"
+    python: "3"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
