+++ docs/conf.py
@@ -14,7 +14,9 @@ release = 'v1.0.0'
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = []
+extensions = [
+    'sphinx_rtd_theme',
+]
 
 templates_path = ['_templates']
 source_suffix = '.rst'
