+++ .readthedocs.yaml
@@ -8,14 +8,14 @@ version: 2
 build:
   os: ubuntu-22.04
   tools:
-    python: "3"
+    python: "3.8"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
     # golang: "1.20"
-    #jobs:
-     # post_create_environment:
-      #  - python -m pip install sphinx-rtd-theme
+    jobs:
+      post_create_environment:
+        - python -m pip install sphinx-rtd-theme
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
