+++ .readthedocs.yaml
@@ -13,9 +13,6 @@ build:
     # nodejs: "20"
     # rust: "1.70"
     # golang: "1.20"
-    jobs:
-      post_create_environment:
-        - python -m pip install sphinx-rtd-theme
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
