### 2024-10-05 954cbc2
Update .readthedocs.yaml
