+++ .readthedocs.yaml
@@ -14,6 +14,12 @@ build:
     # rust: "1.70"
     # golang: "1.20"
 
+python:
+  install:
+    - requirements: docs/requirements.txt
+    - method: pip
+      path: .
+
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/conf.py
