+++ docs/requirements.txt
@@ -1,1 +1,4 @@
-pip install sphinx_rtd_theme 
+Sphinx==1.7.0
+
+# Theme
+sphinx-rtd-theme==0.2.4
