+++ .readthedocs.yaml
@@ -17,8 +17,6 @@ build:
 python:
   install:
     - requirements: docs/requirements.txt
-    - method: pip
-      path: .
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
