+++ docs/requirements.txt
@@ -1,4 +1,3 @@
-Sphinx==1.7.0
-
-# Theme
-sphinx-rtd-theme==0.2.4
+sphinx==5.3.0
+sphinx_rtd_theme==1.1.1
+sphinx-notfound-page==1.0.2
