### 2024-10-05 fa0c1b5
Update .readthedocs.yaml
