+++ .readthedocs.yaml
@@ -6,11 +6,11 @@ version: 2
 
 # Set the OS, Python version and other tools you might need
 build:
-  os: ubuntu-22.04
+  os: ubuntu-24.04
   tools:
-    python: "3.8"
+    python: "3.12"
     # You can also specify other tool versions:
-    # nodejs: "20"
+    nodejs: "20"
     # rust: "1.70"
     # golang: "1.20"
 
