+++ .readthedocs.yaml
@@ -6,19 +6,17 @@ version: 2
 
 # Set the OS, Python version and other tools you might need
 build:
-  os: ubuntu-22.04
+  os: ubuntu-24.04
   tools:
     python: "3.12"
-    nodejs: "18"
-    rust: "1.64"
-    golang: "1.19"
     # You can also specify other tool versions:
-    # nodejs: "20"
+    nodejs: "20"
     # rust: "1.70"
     # golang: "1.20"
-    jobs:
-      post_create_environment:
-        - python -m pip install sphinx-rtd-theme
+
+python:
+  install:
+    - requirements: docs/requirements.txt
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
