+++ docs/index.rst
@@ -24,6 +24,13 @@ Here is some new content from VS code.
 .. toctree::
    :maxdepth: 2
    :hidden:
+   :caption: Contents:
+
+   Contents/API Documentation
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
    :caption: Options:
 
    Options/justlogic
