+++ docs/Contents/APIdocumentation.rst
@@ -1,2 +1,5 @@
 API Documentation
 =================
+
+Functions
+---------
