+++ docs/Contents/APIdocumentation.rst
@@ -3,3 +3,15 @@ API Documentation
 
 Functions
 ---------
+.. list-table:: Comparison
+    :widths: 20 10 10 15
+    :header-rows: 1
+    
+    * - Platform
+      - self
+      - Cost
+      - flex
+    * - rasberry
+      - no
+      - $30
+      - limitless
