+++ docs/index.rst
@@ -7,6 +7,7 @@ COBRAPRO documentation
 ======================
 This is the documentation for CORBAPRO. this is a link :doc:`/Options/justcode`
 this is a link to a particular section :ref:`Options section for justcode <codeOptions>`
+This is the API documentation link :doc:`/Contents/APIdocumentation`
 
 Here is some new content from VS code.
 
