+++ docs/Contents/APIdocumentation.rst
@@ -12,4 +12,13 @@ Functions
    :param param: A structure containing the model parameters.
    :type param: struct
 
-   :returns: * **x_tot** (*casadi.SX or double*) -- A vector where each element corresponds to the solution of the algebraic residual equations g(t,x)=0.
+   :returns: **x_tot** (*casadi.SX or double*) -- A vector where each element corresponds to the solution of the algebraic residual equations g(t,x)=0.
+
+.. py:function:: defineVariableIndex(param)
+
+   Defines the indices and calculates the length for the differential and algebraic variables and stores the information in param structure.
+
+   :param param: A structure containing the model parameters.
+   :type param: struct
+
+   :returns: **param** (*struct*) -- A structure containing the model parameters, now inlcuding the variable lengths and indices.
