+++ docs/Contents/APIdocumentation.rst
@@ -26,11 +26,10 @@ Functions
 .. py:function:: diff_res(t, x, xp, ida_user_data)
 
    This function calculates the differential-algebraic equations (DAEs) in implicit form, depending on the method chosen for determining consistent initial conditions:
-   - If using the SUNDIALS IDACalcIC method (`param.init_Method = 'IDACalcIC'`), it computes the DAE system \( F(t, x, x_p) = 0 \), which includes both the ODE and algebraic equation residuals.
-   - If using the single-step approach [1] (`param.init_Method = 'SS'`), it calculates the implicit ODE system \( M(t, x, x_p) = 0 \) [2], consisting of the perturbed algebraic equations and original ODEs, with the switch function applied.
+   * If using the SUNDIALS IDACalcIC method (`param.init_Method = 'IDACalcIC'`), it computes the DAE system \( F(t, x, x_p) = 0 \), which includes both the ODE and algebraic equation residuals.
+   * If using the single-step approach [1] (`param.init_Method = 'SS'`), it calculates the implicit ODE system \( M(t, x, x_p) = 0 \) [2], consisting of the perturbed algebraic equations and original ODEs, with the switch function applied.
 
-   References
-   ----------
+   **References**:
    [1] M. T. Lawder, V. Ramadesigan, B. Suthar, and V. R. Subramanian, “Extending explicit and linearly implicit ODE solvers for index-1 DAEs,” Computers & Chemical Engineering, vol. 82, pp. 283–292, Nov. 2015, doi: 10.1016/j.compchemeng.2015.07.002.
    [2] S. Ha and S. Onori, “COBRAPRO: An Open-Source Software for the Doyle-Fuller-Newman Model with Co-Simulation Parameter Optimization Framework,” J. Electrochem. Soc., vol. 171, no. 9, p. 090522, Sep. 2024, doi: 10.1149/1945-7111/ad7292.
 
@@ -43,4 +42,6 @@ Functions
    :param ida_user_data: Structure containing additional functions or parameters for the IDA solver.
    :type ida_user_data: struct
 
-   :returns: **param** (*struct*) -- Structure containing model parameters, now updated with variable lengths and indices.
+   :returns: * **dx_tot** (*struct*) -- Structure containing model parameters, now updated with variable lengths and indices.
+             * **flag** (*double*) -- Required by IDA solver but not used in the code.
+             * **new_data** (*double*) -- Required by IDA solver but not used in the code.
