### 2024-10-18 f9e7a6d
Delete docs/Guidelines directory
