+++ docs/Guidelines/justlogic.rst
@@ -1,17 +1,0 @@
-Just Logic
-++++++++++
-
-Key Goals - test
-=================
-
-Options
-=======
-
-CodeMaster
-----------
-
-Robot Turtles
---------------
-
-Primo / Cubetto
----------------
