+++ docs/Guidelines/justcode.rst
@@ -1,6 +1,0 @@
-justcode
-========
-
-Key Goals
----------
-
