### 2024-10-18 2ea5a32
Delete docs/Options directory
