### 2024-10-21 8f15602
Update cs_surf_hermite_interp.m
