### 2024-10-21 9bc6bcf
Update cs_surf_hermite_interp.m
