### 2024-10-22 #6803 @mbarzegary
@COBRAPROsimulator can you please provide an update of your progress on fixing the issues?
