+++ paper/paper.md
@@ -67,11 +67,11 @@ In contrast, several open-source DFN model simulation tools have emerged, such a
 
 # Example: Case Study on LG 21700-M50T Cells
 
-To demonstrate the parameter identification process in COBRAPRO, a case study is conducted to parameterize a fresh LG 21700-M50T cell using the C/20 capacity test, HPPC, and driving cycle data [@pozzato_data_2022]. In this example, we break down the identification problem by systematically grouping parameters in each identification step, as shown in \autoref{fig:flowchart}. This multi-step approach is proposed to improve the identifiability of parameters instead of identifying all the unknown parameters simultaneously [@arunachalam_full_2019]. For detailed information on the multi-objective function and the optimization problem formulated at each identification step, refer to [@ha_cobrapro_2024].
+To demonstrate the parameter identification process in COBRAPRO, a case study is conducted to parameterize a fresh LG 21700-M50T cell using the C/20 capacity test, HPPC, and driving cycle data [@pozzato_data_2022]. In this example, we break down the identification problem by systematically grouping parameters in each identification step, as shown in \autoref{fig:flowchart}. This multi-step approach is proposed to improve the identifiability of parameters instead of identifying all the unknown parameters simultaneously [@arunachalam_full_2019]. For detailed information on the multi-objective cost functions, identifiability analysis, and list of all the DFN parameters, refer to [@ha_cobrapro_2024]. 
 
 ![Case study: Parameter identification procedure on LG 21700-M50T cells.\label{fig:flowchart}](example_flowchart.pdf){ width=100% }
 
-Based the analysis from [@ha_cobrapro_2024], the geometric parameters, OCPs, transference number, and equilibrium electrolyte coefficients are determined to be redundant parameters from structural identifiability analysis. Their values are fixed to values measured from cell tear-down and half-cell experiments on LG 21700-M50 cells, as reported by [@chen_development_2020]. Next, the C/20 capacity test data is used to identify the stoichiometric parameters in the example code `DFN_pso_0_05C.m`. A parameter identifiability study is then performed in `DFN_LSA_Corr_HPPC.m`, which includes LSA and correlation analysis to select parameters that exhibit high sensitivity to HPPC voltage and SOC, while maintaining low correlation with other parameters. The identified subset of parameters is then optimized using HPPC data in the `DFN_pso_HPPC.m` code. Finally, the identified parameters are validated with urban dynamometer driving schedule (UDDS) data in the `DFN_UDDS_validation.m` code. All case study codes are located in the Examples folder.
+Based the analysis from [@ha_cobrapro_2024], the geometric parameters, OCPs, transference number, and equilibrium electrolyte coefficients are determined to be redundant parameters from structural identifiability analysis. Their values are fixed to values measured from cell tear-down and half-cell experiments on LG 21700-M50 cells, as reported by [@chen_development_2020]. Next, the C/20 capacity test data is used to identify the stoichiometric parameters in the example code `DFN_pso_0_05C.m`. A parameter identifiability study is then performed in `DFN_LSA_Corr_HPPC.m`, which includes LSA and correlation analysis to select parameters that exhibit high sensitivity to HPPC voltage and SOC, while maintaining low correlation with other parameters. The identified subset of parameters is then optimized using HPPC data in the `DFN_pso_HPPC.m` code. Finally, the identified parameters are validated with urban dynamometer driving schedule (UDDS) data in the `DFN_UDDS_validation.m` code. All case study codes are located in the Examples folder. 
 
 View COBRAPRO's [README on Github](https://github.com/COBRAPROsimulator/COBRAPRO/blob/main/README.md#toc6) to view a list of all available example codes.
 
