### 2024-10-31 #6803 @COBRAPROsimulator
Hello @brosaplanella, thank you again for your valuable suggestions. We have made changes to our code and paper accordingly. In addition to our responses to your comments in the thread above, please find the actions we took for each of your comments below. Please let us know any further clarification you require and other suggestions/comments you may have. Thank you very much.

**Major comments**
- One of the main claims in the paper is performance, and it is stated that the 1C discharge runs in less than a second (similar order of magnitude to PyBaMM). When running it on my laptop (it tic/toced the first 1C discharge in cycle_CC.m it took around 8 seconds (the PyBaMM time in my machine is similar to that reported by the authors). I understand that the discrepancy is probably due to pre/post-processing and printing to screen, but it would be good to have one example with minimal overhead which can demonstrate the performance claimed in the article.
- [x] Actions taken: In line 344 of the DFN_model/runModel.m function, the script will now print the time it took to solve the DAE to the Command Window. This way, if the printing option is enabled, users will be able to readily view the DAE solver time without having to resort to tic/toc. Furthermore, in Examples/Cycling/cycle_CC.m, lines 141-148 now prints the DAE solver time for each cycle and the user defined total number of cycles. For 1C discharge with Np=Ns=Nn=Nrp=Nrn=10, it can be verified with cycle_CC.m that it takes less than 1 second to run one cycle, as claimed in the paper. Also, please refer to our JES paper for a comprehensive computation time comparison study: S. Ha and S. Onori, “COBRAPRO: An Open-Source Software for the Doyle-Fuller-Newman Model with Co-Simulation Parameter Optimization Framework,” J. Electrochem. Soc., vol. 171, no. 9, p. 090522, Sep. 2024, doi: [10.1149/1945-7111/ad7292](https://doi.org/10.1149/1945-7111/ad7292).

- I agree with [@yuefan98](https://github.com/yuefan98) that the documentation is a bit sparse. Even though for a research tool the current documentation might be enough, it should be expanded in the near future (I am also happy for this to take place after publication). I believe the main issue is that, even though the examples work straight out of the box, it would be extremely hard for a user to adapt the code for their own needs. I believe that publishing some MATLAB notebooks commenting on the steps would be very useful.
- [x] Actions taken: A [readthedocs](https://cobrapro.readthedocs.io/en/latest/Contents/APIdocumentation.html) have been created for COBRAPRO. This includes the API documentation listing the functions in COBRAPRO. The authors plan to expand upon readthedocs to include examples/tutorials to allow for a more intuitive user experience. The readthedocs is still a work in progress and it is now linked on the README page.

- Similarly, tests are quite sparse. At the moment the tests only check that CasADI runs and that one particular example works, but these need to be ran manually. I would recommend some continuous integration on Github (or similar) to test things work periodically, but I don't know if that is feasible with proprietary software like MATLAB. In any case, some additional tests should be written, at the very least one to check that the parameter fitting tools work as expected.
- [x] Actions taken: We have included an additional test,  test/test3_psoCheck.m, that checks that the particle swarm optimization function is working in the parallel structure, as required by the parameter optimization toolbox in COBRAPRO.

- At the moment the list of CONTRIBUTORS is a bit opaque. All the commits have been pushed by COBRAPROsimulator, which does not tell who that is or if is one or multiple users. This could be an issue if other people join the project or the person behind COBRAPROsimulator changes institutions. I would recommend having a list of contributors in the README, where contributors are acknowledged. The [all-contributors](https://allcontributors.org/) tool is quite useful for this.
- [ ] The authors will use the all-contributors tool to list all the contributors in the README page. We are still in the process of setting this up, but wanted to inform you that we have taken action on your advice.

Minor comments
- The instructions require SUNDIALS 2.6.2 but the hyperlink points to 2.6.1. In addition, this version is quite old. Is there any particular reason to not use a more recent version?
- [x] Actions taken: The link has been fixed to SUNDIALS 2.6.2. To add to my response above on why the older version of SUNDIALS is used, please refer to our JES paper that explains this further: S. Ha and S. Onori, “COBRAPRO: An Open-Source Software for the Doyle-Fuller-Newman Model with Co-Simulation Parameter Optimization Framework,” J. Electrochem. Soc., vol. 171, no. 9, p. 090522, Sep. 2024, doi: [10.1149/1945-7111/ad7292](https://doi.org/10.1149/1945-7111/ad7292).

- The instructions say CasADI should be installed, but they don't specify for which platform. Even though it is quite obvious, it would be helpful to specify it needs to be the MATLAB version.
- [x] Actions taken: In the installation directions of the README file, it has been specified that the MATLAB version of CasADi needs to be installed.

- The instructions assume MATLAB is installed, with some toolboxes. It should be stated explicitly at the beginning of the "Installation" section.
- [x] Actions taken: The installation directions of the README file has been edited such that the first step is to download MATLAB and the necessary MATLAB toolboxes.

- There is a step that asks "Install toolbox" that it is not reflected in the installation instructions.
- [x] Actions taken: The “Install toolbox” step has been added to the installation directions in the README file.

- It would be good to always have a default answer to most of the prompts (so just hitting enter gets you through), as the fact that some answers require a yes and some a no is a bit confusing. If that is already implemented, users should be made aware of what the default is.
- [x] Actions taken: Since the yes/no answers are embedded as part of the SUNDIALS installation function, it seems that there is no way to edit or change the answers. Furthermore, the yes/no answers allow users to download specific parts of the SUNDIALS package, which provides users with more flexibility.

- The title of the correlation matrix plot (e.g. when running DFN_LSA_Corr_CC.m) does not render LaTeX.
- [x] Actions taken: It is now verified that all LaTeX renderings in DFN_LSA_Corr_CC.m work as expected.

- The DFN_pso_0_05C.m also took a long time to me (around 30 minutes) and some errors/warning came up. It would be good to clarify in the documentation that this example might take a while to run.
- [x] Actions taken: It has been noted in the README file that the parameter identification example files take a while to run. Also, the files parameter_identification/pso_objectiveFunctions/DFN_obj_CC.m and parameter_identification/pso_objectiveFunctions/DFN_obj_HPPC.m have been edited such that the messages printed to the Command Window are clearer to understand. The message now prints what occurred in the objective function, such as ‘'Model failed to initialize algebraic variables for given PSO parameter set” or “'PSO caught unfeasible solution” rather than printing “errors” which can be confusing to the users, as these are not errors but rather instances of the DFN model unable to be solved given the parameter guesses by PSO.

- [l33] I would remove all the references here, as they are all cited again and discussed in more detail later on.
- [x] Actions taken: The reference has been removed from [l33].

- [l47] "Other packages resort to literature-derived parameters and lack the ability to predict battery behaviour". I think this statement is a bit of a stretch. All the aforementioned tools will provide models with the ability to predict real battery behaviour, as long as the user provides the right parameter values. COBRAPRO provides tools to estimate those parameters (which is no small feat), while the others don't, so I think this would be a more accurate statement.
- [x] Actions taken: This part of the paper has been modified to clarify that COBRPRO allows the determination of parameters (closed-loop software), whereas other open-source DFN codes require pre-determined parameters (open-loop software) to simulate real-world batteries.

- [l54]: "COBRAPRO leverages a fast solver". Which one? Give details and references.
- [x] Actions taken: The paper has been edited to specify that the SUNDIALS IDA solver is used and relevant references have been added as well.

- [l64] About Challenge 2, I think it is a bit misleading. I suspect most of the other tools also look for consistent initial conditions (PyBaMM definitely does, and I suspect COMSOL does too; haven't played much with the other tools). If any of the other tools does not do that, it should be stated clearly. Otherwise, I think this challenge would be better merged with Challenge 1, as it all ties to dealing with the computational complexity.
- [x] Actions taken: Challenge 2 has been rewritten to highlight and distinguish the consistent initial condition method used in COBRAPRO and other open-source DFN codes. COBRAPRO uses the single-step approach, which perturbs algebraic equations into implicit ordinary differential equations, preventing the need for a nonlinear algebraic solver like SUNDIALS IDACalcIC or CasADi rootfinder. Nonlinear algebraic solver are typically used in other open-source DFN codes as highlighted in our JES paper, which has been shown to be less robust than the single-step method at determining consistent initial conditions at various discretization points and C-rates.

- [l185, but there are other occurrences through the text] The subscripts denoting labels (e.g. n, p, e...) should be in Roman text (i.e. \mathrm{e}) in LaTeX.
- [x] Actions taken: Authors note that the variables in [l185] (now [l158] in the new version) refer to the variable names in the code, not the LaTeX versions of the parameter names.

- [Fig3&5] The plots showing the SOC are not very insightful. I imagine they are computed through Coulomb counting in both cases, and given that the experiment is current driven they need to be virtually identical. I believe a plot of the voltage error between experiment and simulation would be a more valuable output of the code.
- [x] Actions taken: In [l93] of the new version, authors have added more details and cited the multi-objective function proposed by [A. Allam and S. Onori, “Online Capacity Estimation for Lithium-Ion Battery Cells via an Electrochemical Model-Based Adaptive Interconnected Observer,” IEEE Trans. Contr. Syst. Technol., vol. 29, no. 4, pp. 1636–1651, Jul. 2021, doi: [10.1109/TCST.2020.3017566](https://doi.org/10.1109/TCST.2020.3017566)]. This highlights the reason for including the SOC plots in the identification results.

- [l246] Given that this list is bound to change as the code evolves, I think it would be better to not include it in the article and, instead, point the users to the README where an up-to-date list will be available.
- [x] Actions taken: The example files and location have been removed from the paper. The authors plan to include update the readthedocs accordingly as more examples are included in the future.
