### 2024-11-05 #6803 @COBRAPROsimulator
Hello @BradyPlanden, thank you again for your helpful comments. We have made changes to our code and paper to reflect your comments. Below, please find the action items we took for each of your comments. Please let us know if you require clarification on any items discussed.

Repository Comments
- While the authors provide an initial testing framework, this only covers conventional integration tests and lacks individual unit tests or providing users with a coverage metric.

- [ x] Actions taken: We have included an additional test (test/test3_psoCheck.m) that checks for correct implementation of the particle swarm optimization function in parallel structure, as required by the parameter optimization toolbox in COBRAPRO. By individual unit tests, does this mean providing a way to check that each MATLAB function works correctly? The authors would greatly appreciate additional details/information to improve the testing metrics of COBRAPRO.

- In it's current state, the repository appears to be more of a collection of scripts rather than a package/module/etc. This makes it difficult to determine the software architecture, potentially limiting the reach of this work. This repository provides a reference to a possible structure (as a MATLAB codebase): https://github.com/Battery-Intelligence-Lab/BatEst
- [x] Actions taken: The authors have looked at the BatEst repository and categorized COBRAPRO in a similar manner such that each folder and subfolders are more intuitively organized. For instance, the original COBRAPRO folders were organized as “DFN_identification”  (included PSO related functions), “DFN_identification_results” (included identified results in .mat files), “DFN_model” (included all DFN model related functions), “DFN_parameters” (included DFN parameter files), “Local_Sensitivity Analysis”, “Examples”, and “test”. Now, the folders are organized as “data” (includes battery data used in our JES paper), “parameters” folder (consisting of subfolders for OCP and electrolyte-related functions), “parameter_identification” (consisting of subfolders “identifiability_analysis” and “pso_objectiveFunctions”, and “DFN_model” (consisting of subfolders “additional equations”, “DAE”, FVM_interpolation”, and “helper_functions”). The “parameters” folder is organized in this manner such that the functional parameters (OCP and electrolyte-related parameters) can be saved in their respective folders. The “parameter_identification” folder includes the closed-loop PSO function and parameter identifiability analysis (LSA and correlation analysis), rather than having them in separate main folders. The “DFN_model” subfolders are organized according to the code’s functions, such as “DAE” for function related to setting up the DAE, “FVM_interpolation” for numerical discretization related functions, “additional_equations” for functions to calculate the particle concentration related equations once the DAE is solved, and “helper_functions” for accessory functions to aid the main function, runModel.m.

- More functionality could be added to the particle swarm optimisation object (i.e. max iterations, time, etc.) with documentation. Specifically, here: https://github.com/COBRAPROsimulator/COBRAPRO/blob/157ab80ffda30aa7c51c39cc7a042162efbfbaf0/Examples/Parameter_Identification_Routines/DFN_pso_0_05C.m#L198
- [x] Actions taken: The files Examples/Parameter_Identification_Routines/DFN_pso_0_05C.m and Examples/Parameter_Identification_Routines/DFN_pso_HPPC.m have been edited to allow users to modify the PSO settings. For DFN_pso_0_05C.m, lines 83-87 can be edited to change the particle number, lines 89-106 allow users to determine the exit conditions in the PSO, and lines 108-116 can be edited to change the weight settings in PSO (appropriate references have been added for users to learn more about PSO weight settings). Similarly, DFN_pso_HPPC.m has been edited to include the same new features in lines 90-123.

Article Comments
- Presently, the comparison with alternative work in the field is aimed at pure forward modelling packages (PyBaMM, PetLion.jl, LionSimba) which do not aim to solve the inverse problem, a more extensive literature review will find other implementations which are more suitable for comparison.
- Actions taken: In lines 38-46 of the paper, authors have specified that the comparison is made between open-source codes for the DFN model. While other open-source codes may be available for the SPM, ESPM, and ECM that include parameter identification routines, we highlight that there is a lack of parameter identification codes for the DFN model other than DEARLIBS. The DFN model is the most computationally complex and highest fidelity model compared to the reduced-order forms and the empirical ECM. For this reason, the DFN model provides unique advantages for material design and advanced BMS applications, highlighting the need for an open-source DFN code that can determine accurate parameters through parameter identification. To further clarify, we do not solve the inverse problem (we do not inverse the dynamic equations), but rather we are using an iterative optimization process to determine parameters using battery voltage and current data. Please let us know if we can clarify any other points.

- A table showing the full set of parameters identified for the example would be helpful to summarise the results in one place for the reader.
- [x] Actions taken: To adhere to JOSS’s paper length, authors have included in line 133 a reference to our JES paper, which includes a comprehensive table listing all the DFN model parameters.

- Similarly, a table of correlation metrics for the parameters in the LSA section of the example.
- [x] Actions taken: Authors have rewritten the LSA section in the example to clarify how the identifiability analysis works. Now, this part is titled “Parameter Identifiability Analysis” and shows how to determine an identifiable parameter set using LSA and correlation analysis. The user defines a correlation threshold, which is then used to determine the if any of two parameters are correlated or not. The algorithm behind this is explained in detail in our JES paper, S. Ha and S. Onori, “COBRAPRO: An Open-Source Software for the Doyle-Fuller-Newman Model with Co-Simulation Parameter Optimization Framework,” J. Electrochem. Soc., vol. 171, no. 9, p. 090522, Sep. 2024, doi: [10.1149/1945-7111/ad7292](https://doi.org/10.1149/1945-7111/ad7292), which is cited in this section. The correlation threshold is the only user defined metric for this example, as described in the new version of the paper in lines 204-208.
