### 2024-11-15 #6803 @BradyPlanden
Yes - thanks for the update @COBRAPROsimulator. Happy for this to proceed @mbarzegary.
