### 2024-11-18 e6d64bc
fix reference citation rendering issue
