### 2024-11-18 04c9f29
fix typos in joss paper
