+++ paper/paper.md
@@ -15,16 +15,16 @@ authors:
     orcid: 0000-0002-6556-2608
     affiliation: 2
 affiliations:
- - name: Mechanical Engineering, Stanford University, 440 Escondido Mall, Stanford 94305, CA, USA
+ - name: Mechanical Engineering, Stanford University, 440 Escondido Mall, Stanford 94305, CA, United States of America
    index: 1
- - name: Energy Science and Engineering, Stanford University, 367 Panama Mall, Stanford 94305, CA, USA
+ - name: Energy Science and Engineering, Stanford University, 367 Panama Mall, Stanford 94305, CA, United States of America
    index: 2
 date: 01 March 2024
 bibliography: paper.bib
 ---
 
 # Summary
-COBRAPRO (**Co**-simulation **B**atte**r**y Modeling for **A**ccelerated **P**a**r**ameter **O**ptimization) is a physics-based battery modeling software with the capability to perform closed-loop parameter optimization using experimental data. COBRAPRO is based on the Doyle-Fuller-Newman (DFN) model [@doyle_modeling_1993], which is the most widely-accepted high-fidelity model that considers the lithium-ion transport and charge conservation in the liquid electrolyte and solid electrodes, and kinetics at the solid and liquid interface during lithium intercalation and deintercalation. Such physics-based models have found applications in battery design [@dai_graded_2016; @couto_lithiumion_2023] and advanced battery management systems to ensure reliable and safe operation of electric vehicles [@kolluri_realtime_2020]. The DFN model relies on physical parameters such as geometric, stoichiometric, concentration, transport, and kinetic properties. Accurate parameter values are essential for the model to predict real battery behavior under various usage conditions. However, parameters obtained from cell teardown experiments cannot be directly applied [@ha_cobrapro_2024; @li_datadriven_2022], as the DFN model simplifies a real battery, assuming perfectly spherical particles, neglecting electrode heterogeneity, and considering internal dynamics only in one dimension. Parameter identification is therefore crucial for developing models that accurately capture battery dynamics. COBRAPRO enables users to identify these parameters for any battery using standard current-voltage data from a battery cycler, optimizing DFN parameters by minimizing the error between simulated and experimental results through an integrated optimization routine.
[…]
