### 2024-11-18 #6803 @mbarzegary
@COBRAPROsimulator given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please
* Merge my PR containing some minor edits
* Work on the author's points of the final checklist I created ☝️
