### 2024-11-20 #6803 @COBRAPROsimulator
Hi @mbarzegary, thank you very much for initiating the post-review process. I have completed the post-review tasks as you mentioned:

- [x ] I have double checked the authors and affiliations (including ORCIDs)
- [ x] The version of the software that will be used in the JOSS paper is v2.0.0 (includes all changes made from the review process).
- [x] I have created a Zenodo archive and the DOI is: 10.5281/zenodo.14192734
- [x] I made sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
- [x] I made sure that the license listed for the archive is the same as the software license.
