### 2024-11-25 #6803 @mbarzegary
@COBRAPROsimulator Nice. The only issues I see is that the files in the archive are uploaded one by one, meaning that the directory structure of the repository is lost. Can you please instead upload the zip file of the repository so that the directory structure retains? An example would this: https://zenodo.org/records/14052179
