### 2024-11-25 #6803 @COBRAPROsimulator
Hi @mbarzegary, thank you for pointing this out. The new Zenodo DOI is: 10.5281/zenodo.14192733
