### 2024-12-05 #6803 @xuanxu
The entry `xu_comparative_2023` is duplicated in the .bib file
