### 2024-12-07 3c601c0
Update paper.bib
