### 2024-12-07 #6803 @COBRAPROsimulator
@xuanxu, thank you for pointing out the issue. @mbarzegary I have removed the duplicated entry in the bib file.
