### 2024-12-11 #6803 @kyleniemeyer
@COBRAPROsimulator just one minor issue in the paper to resolve: the COMSOL reference appears incorrectly ("COMSOL multiphysics&copy;"). Can you please fix this?
