+++ paper/paper.bib
@@ -251,14 +251,14 @@
 
 @misc{comsol,
   author = {{COMSOL AB, Stockholm, Sweden}},
-  title = {COMSOL Multiphysics&copy; v. 6.2},
+  title = {COMSOL Multiphysics{&copy}; v. 6.2},
   url = {www.comsol.com},
   year = {2023},
 }
 
 @misc{comsol_battery,
   author = {{COMSOL AB, Stockholm, Sweden}},
-  title = {Battery Design Module User's Guide, COMSOL Multiphysics&copy; v. 6.2},
+  title = {Battery Design Module User's Guide, COMSOL Multiphysics{&copy}; v. 6.2},
   year = {2023},
 }
 
