+++ paper/paper.bib
@@ -251,7 +251,7 @@
 
 @misc{comsol,
   author = {{COMSOL AB, Stockholm, Sweden}},
-  title = {COMSOL Multiphysics{&copy}; v. 6.2},
+  title = {COMSOL Multiphysics; v. 6.2},
   url = {www.comsol.com},
   year = {2023},
 }
