+++ paper/paper.bib
@@ -256,12 +256,6 @@
   year = {2023},
 }
 
-@misc{comsol_battery,
-  author = {{COMSOL AB, Stockholm, Sweden}},
-  title = {Battery Design Module User's Guide, COMSOL Multiphysics{&copy}; v. 6.2},
-  year = {2023},
-}
-
 @misc{fastDFN,
   author = {{Scott Moura}},
   title = {fastDFN},
