### 2024-12-12 4b8b18c
Update paper.bib
