+++ paper/paper.bib
@@ -249,9 +249,9 @@
   doi = {10.1149/1945-7111/ad1293},
 }
 
-@misc{comsol,
+@article{comsol,
   author = {{COMSOL AB, Stockholm, Sweden}},
-  title = {COMSOL Multiphysics {\textregistered}; v. 6.2},
+  title = {COMSOL Multiphysics{\textregistered}; v. 6.2.},
   url = {www.comsol.com},
   year = {2023},
 }
