+++ paper/paper.bib
@@ -251,7 +251,7 @@
 
 @article{comsol,
   author = {{COMSOL AB, Stockholm, Sweden}},
-  title = {COMSOL Multiphysics{\textregistered}; v. 6.2.},
+  title = {COMSOL {Multiphysics}{\textregistered}; v. 6.2.},
   url = {www.comsol.com},
   year = {2023},
 }
