+++ paper/paper.md
@@ -30,7 +30,7 @@ COBRAPRO (**Co**-simulation **B**atte**r**y Modeling for **A**ccelerated **P**a*
 
 Although parameter calibration is required to accurately predict the dynamical behavior of real batteries, current DFN modeling tools lack the capability to perform parameter identification. 
 
-COMSOL Multiphysics&copy; [@comsol] is a commercially available finite element modeling software commonly used to simulate the DFN model. Although COMSOL lacks a built-in parameter identification feature, it was demonstrated that COMSOL's *LiveLink&trade; for MATLAB*&copy; can be used to establish communication between COMSOL and MATLAB for parameter optimization  [@pozzato_general_2023]. This framework allows users to leverage the versatile suite of optimizers in MATLAB while running COMSOL to generate the model output. However, the expensive licensing fee and proprietary nature of COMSOL create barriers to public access, limiting collaboration and code reproducibility.
+COMSOL Multiphysics^&copy^; [@comsol] is a commercially available finite element modeling software commonly used to simulate the DFN model. Although COMSOL lacks a built-in parameter identification feature, it was demonstrated that COMSOL's *LiveLink&trade; for MATLAB*&copy; can be used to establish communication between COMSOL and MATLAB for parameter optimization  [@pozzato_general_2023]. This framework allows users to leverage the versatile suite of optimizers in MATLAB while running COMSOL to generate the model output. However, the expensive licensing fee and proprietary nature of COMSOL create barriers to public access, limiting collaboration and code reproducibility.
 
 In contrast, several open-source DFN model simulation tools have emerged, such as PyBaMM [@sulzer_python_2021], LIONSIMBA [@torchio_lionsimba_2016], PETLION [@berliner_methods_2021], DEARLIBS [@lee_robust_2021], fastDFN [@fastDFN], and MPET [@smith_multiphase_2017]. Among these packages, DEARLIBS is the only software equipped with a closed-loop parameter identification routine. Other open-source packages operate as open-loop software, requiring predefined parameters to simulate real battery behavior. Without parameter optimization capabilities, these open-loop tools typically depend on uncalibrated, experimentally measured parameters from the literature. Taking inspiration from DEARLIBS, COBRAPRO aims to address three primary challenges in the DFN model:
 
