+++ paper/paper.bib
@@ -24,7 +24,7 @@
 }
 
 @article{couto_lithiumion_2023,
-  title = {Lithium-Ion Battery Design Optimization Based on a Dimensionless Reduced-Order Electrochemical Model},
+  title = {{Lithium-Ion Battery Design Optimization Based on a Dimensionless Reduced-Order Electrochemical Model}},
   author = {Couto, {\relax Luis}. D. and Charkhgard, Mohammad and Karaman, Berke and Job, Nathalie and Kinnaert, Michel},
   year = {2023},
   month = jan,
@@ -36,7 +36,7 @@
 }
 
 @article{kolluri_realtime_2020,
-  title = {Real-Time {{Nonlinear Model Predictive Control}} ({{NMPC}}) {{Strategies}} Using {{Physics-Based Models}} for {{Advanced Lithium-ion Battery Management System}} ({{BMS}})},
+  title = {{Real-Time} {{Nonlinear Model Predictive Control}} ({{NMPC}}) {{Strategies}} Using {{Physics-Based Models}} for {{Advanced Lithium-ion Battery Management System}} ({{BMS}})},
   author = {Kolluri, Suryanarayana and Aduru, Sai Varun and Pathak, Manan and Braatz, Richard D. and Subramanian, Venkat R.},
   year = {2020},
   month = jan,
@@ -127,7 +127,7 @@
 }
 
 @article{harris_failure_2017,
-  title = {Failure Statistics for Commercial Lithium Ion Batteries: {{A}} Study of 24 Pouch Cells},
+  title = {{Failure Statistics for Commercial Lithium Ion Batteries: {{A}} Study of 24 Pouch Cells}},
   shorttitle = {Failure Statistics for Commercial Lithium Ion Batteries},
   author = {Harris, Stephen J. and Harris, David J. and Li, Chen},
   year = {2017},
@@ -151,7 +151,7 @@
 }
 
 @article{pozzato_data_2022,
-  title = {Lithium-Ion Battery Aging Dataset Based on Electric Vehicle Real-Driving Profiles},
+  title = {{Lithium-Ion Battery Aging Dataset Based on Electric Vehicle Real-Driving Profiles}},
   author = {Pozzato, Gabriele and Allam, Anirudh and Onori, Simona},
   year = {2022},
   month = apr,
@@ -177,7 +177,7 @@
 }
 
 @article{ha_cobrapro_2024,
-  title = {COBRAPRO: An Open-source Software for the Dolye-Fuller-Newman Model with Co-simulation Parameter Optimization Framework},
+  title = {{COBRAPRO: An Open-source Software for the Dolye-Fuller-Newman Model with Co-simulation Parameter Optimization Framework}},
   author = {Ha, Sara and Onori, Simona},
   journal = {Journal of The Electrochemical Society},
   year = {2024},
@@ -212,7 +212,7 @@
 }
 
 @article{methekar_perturbation_2011,
-  title = {A Perturbation Approach for Consistent Initialization of Index-1 Explicit Differential{\textendash}Algebraic Equations Arising from Battery Model Simulations},
+  title = {{A Perturbation Approach for Consistent Initialization of Index-1 Explicit Differential{\textendash}Algebraic Equations Arising from Battery Model Simulations}},
   author = {Methekar, Ravi N. and Ramadesigan, Venkatasailanathan and Pirkle, J. Carl and Subramanian, Venkat R.},
   year = {2011},
   month = nov,
@@ -264,7 +264,7 @@
 }
 
 @article{gardner2022sundials,
-	title     = {Enabling new flexibility in the {SUNDIALS} suite of nonlinear and differential/algebraic equation solvers},
[…]
