### 2024-12-12 #6803 @COBRAPROsimulator
@kyleniemeyer, thank you for pointing that out. I have fixed the COMSOL reference. Please let me know if there any other issues that come up.
