### 2024-12-21 #6803 @COBRAPROsimulator
Hello @kyleniemeyer, is there a status update regarding  our paper? Please let us know if there are any issues that we need to address. Thank you.
