### 2025-01-02 #6803 @kyleniemeyer
Happy new year and congratulations @COBRAPROsimulator on your article's publication in JOSS! Please consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.

Many thanks to @yuefan98, @BradyPlanden, and @brosaplanella for reviewing this submission, and @mbarzegary for editing.
